% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassAlignedConsensusSet.R
\docType{class}
\name{AlignedConsensusRead-class}
\alias{AlignedConsensusRead-class}
\title{AlignedConsensusRead}
\description{
An S4 class for storing multiple single consensus reads to
}
\section{Slots}{

\describe{
\item{\code{consensusReadsList}}{}
}}

\examples{
rawDataDir <- system.file("extdata", package = "sangeranalyseR")
inputFilesParentDir <- file.path(rawDataDir, "Allolobophora_chlorotica")
suffixForwardRegExp <- "_[F]_[0-9]*.ab1"
suffixReverseRegExp <- "_[R]_[0-9]*.ab1"
AlignedConsensusRead <- new("AlignedConsensusRead",
                     parentDirectory       = inputFilesParentDir,
                     suffixForwardRegExp   = suffixForwardRegExp,
                     suffixReverseRegExp   = suffixReverseRegExp,
                     cutoffQualityScore    = 20,
                     slidingWindowSize     = 8)
}
\author{
Kuan-Hao Chao
}
