% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassBaseCall.R
\docType{class}
\name{BaseCall-class}
\alias{BaseCall-class}
\title{BaseCall}
\description{
An S4 class for base call information
}
\section{Slots}{

\describe{
\item{\code{secondaryPeakRatio}}{}
}}

\author{
Kuan-Hao Chao
}
