% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.stop.codons.R
\name{count.stop.codons}
\alias{count.stop.codons}
\title{Count stop codons in a DNA sequence}
\usage{
count.stop.codons(sequence, reading.frame = 1,
  genetic.code = GENETIC_CODE)
}
\arguments{
\item{sequence}{a DNAString object}

\item{reading.frame}{a number from 1 to 3 denoting the reading frame of the sequence}

\item{genetic.code}{Named character vector in the same format as GENETIC_CODE (the default), which represents the standard genetic code. This is the code with which the function will attempt to translate your DNA sequences. You can get an appropriate vector with the getGeneticCode() function. The default is the standard code.}
}
\description{
Count stop codons in a DNA sequence
}
