\name{HIVR-package}
\alias{HIVR-package}
\alias{HIVR}
\docType{package}
\title{
\packageTitle{HIVR}
}
\description{
\packageDescription{HIVR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HIVR}
\packageIndices{HIVR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{HIVR}

Maintainer: \packageMaintainer{HIVR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
