% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{launchAppSC}
\alias{launchAppSC}
\title{Method launchAppSC}
\usage{
launchAppSC(object, outputDir = NULL, colors = "default")
}
\arguments{
\item{object}{A SangerContig S4 instance.}

\item{outputDir}{The output directory of the saved new SangerContig S4 instance.}

\item{colors}{A vector for users to set the colors of (A, T, C, G, else). 
There are three options for users to choose from. 
  1. "default":  (green, blue, black, red, purple). 
  2. "cb_friendly":  ((0, 0, 0), (199, 199, 199), (0, 114, 178), (213, 94, 0), (204, 121, 167)). 
  3. Users can set their own colors with a vector with five elements.}
}
\value{
A \code{shiny.appobj} object.
}
\description{
Method launchAppSC
}
\examples{
data(sangerContigData)
\donttest{
launchAppSC(sangerContigData)}
}
