% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.abi.reads.R
\name{load.readgroups}
\alias{load.readgroups}
\title{Group ABI reads by recursively reading them from an input folder}
\usage{
load.readgroups(input.folder, forward.suffix, reverse.suffix,
  processors = NULL)
}
\description{
Load all reads recursively, then group them by name after removing prefix/suffix. Return a list of lists of filenames of input reads.
}

