% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.reads.R
\name{load.readsets}
\alias{load.readsets}
\title{Automatically load readsets by grouping .ab1 files by name.}
\usage{
load.readsets(input.folder, forward.suffix, reverse.suffix, processors = NULL,
  min.length = NULL, max.secondary.peaks = NULL)
}
\description{
Load all reads recursively, then group them by name after removing prefix/suffix. Return a list of lists of filenames of input reads.
}

