% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.readgroup.R
\name{make.readgroup}
\alias{make.readgroup}
\title{Make a readgroup object}
\usage{
make.readgroup(fwd.fnames, rev.fnames, name = NULL, processors = NULL)
}
\description{
A readgroup object contains a set of reads that all correspond to a single underlying sequence. The function takes two lists of filenames: those for forward reads and those for reverse reads. It then efficiently reads in all of the files, and returns a single object which contains two lists of forward and reverse sequences corresponding to the forward and reverse filenames respectively.
}

