% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.readset.R
\name{make.readset}
\alias{make.readset}
\title{Make a readgroup object}
\usage{
make.readset(fwd.fnames, rev.fnames, trim = TRUE, trim.cutoff = 0.05,
  trim.segment = 20, name = NULL, processors = NULL)
}
\arguments{
\item{fwd.fnames}{a list of full file paths to forward reads from ab1 files (i.e. those that do not need to be reverse-complemented).}

\item{rev.fnames}{a list of full file paths to reverse reads from ab1 files (i.e. those that *do* need to be reverse-complemented).}

\item{trim}{TRUE/FALSE trim sequences based on quality while creating readgroup? If TRUE, the trim.mott function is applied to each sequence before inclusion in the readgroup. Note, trimming only works if the raw data are stored in ab1 files with appropriate information.}

\item{trim.cutoff}{value passed to trim.mott as quality cutoff for sequencing trimming, only used if 'trim' == TRUE}

\item{trim.segment}{value passed to trim.mott as minimum seuqence length for sequence trimming, only used if 'trim' == TRUE}

\item{name}{an optional name for your readgroup}

\item{processors}{The number of processors to use, or NULL (the default) for all available processors

A set of unaligned reads as a DNAstringset object, names are the input file paths.}
}
\description{
Make a readgroup object
}

