% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{MakeBaseCalls}
\alias{MakeBaseCalls}
\title{Method MakeBaseCalls}
\usage{
MakeBaseCalls(object, signalRatioCutoff = 0.33)
}
\arguments{
\item{object}{object}

\item{signalRatioCutoff}{signalRatioCutoff}
}
\description{
Method MakeBaseCalls
}
