% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{class}
\name{numericORNULL-class}
\alias{numericORNULL-class}
\title{#' @export
setGeneric("generateReportSangerContig", function(obj, outputDir = NULL,
                                      navigationContigFN = NULL,
                                      includeSangerRead = TRUE) {
    standardGeneric("generateReportSangerContig")
})}
\description{
#' @export
setGeneric("generateReportSangerRead", function(obj, outputDir = NULL,
                                      navigationContigFN = NULL,
                                      navigationAlignmentFN = NULL) {
    standardGeneric("generateReportSangerRead")
})
}
