% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{preQualityBasePlot}
\alias{preQualityBasePlot}
\title{Method preQualityBasePlot}
\usage{
preQualityBasePlot(object, readFeature = "Forward Read")

preQualityBasePlot(object, readFeature = "Forward Read")
}
\description{
Method preQualityBasePlot

Method preQualityBasePlot
}
