% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{primaryAASeqS3}
\alias{primaryAASeqS3}
\title{Method primaryAASeqS3}
\usage{
primaryAASeqS3(object)
}
\arguments{
\item{object}{A SangerRead S4 instance.}
}
\value{
Frame-3 AA translation as \code{AAString} (lazy).
}
\description{
Method primaryAASeqS3
}
