% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{qualityBasePlot}
\alias{qualityBasePlot}
\title{Method qualityBasePlot}
\usage{
qualityBasePlot(object)
}
\arguments{
\item{object}{object}
}
\description{
Method qualityBasePlot
}
\examples{
data(qualityReport)
data(sangerReadF)
qualityBasePlot(qualityReport)
qualityBasePlot(sangerReadF)
}
