% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsQualityReport.R
\docType{methods}
\name{QualityReport-class-preQualityBasePlot}
\alias{QualityReport-class-preQualityBasePlot}
\alias{preQualityBasePlot,QualityReport-method}
\title{preQualityBasePlot}
\usage{
\S4method{preQualityBasePlot}{QualityReport}(object)
}
\arguments{
\item{object}{object}
}
\description{
A QualityReport method which pre-plots quality base interactive plot.
}
\examples{
\dontrun{preQualityBasePlot(qualityReport)}
}
