% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsQualityReport.R
\name{QualityReport-class-qualityBasePlot}
\alias{QualityReport-class-qualityBasePlot}
\alias{qualityBasePlot,QualityReport-method}
\title{qualityBasePlot}
\usage{
\S4method{qualityBasePlot}{QualityReport}(object)
}
\arguments{
\item{object}{A QualityReport S4 instance.}
}
\value{
A quality plot.
}
\description{
A QualityReport method which creates quality base interactive plot.
}
\examples{
data("qualityReportData")
\donttest{
qualityBasePlot(qualityReportData)}
}
