% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsQualityReport.R
\docType{methods}
\name{QualityReport-class-updateQualityParam}
\alias{QualityReport-class-updateQualityParam}
\alias{updateQualityParam,QualityReport-method}
\title{updateQualityParam}
\usage{
\S4method{updateQualityParam}{QualityReport}(object,
  TrimmingMethod = "M1", M1TrimmingCutoff = 1e-04,
  M2CutoffQualityScore = NULL, M2SlidingWindowSize = NULL)
}
\arguments{
\item{object}{object}

\item{TrimmingMethod}{TrimmingMethod}

\item{M1TrimmingCutoff}{M1TrimmingCutoff}

\item{M2CutoffQualityScore}{M2CutoffQualityScore}

\item{M2SlidingWindowSize}{M2SlidingWindowSize}
}
\description{
A QualityReport method which updates quality base interactive plot.
}
\examples{
\dontrun{updateQualityParam(qualityReport,
                   TrimmingMethod         = "M2",
                   M1TrimmingCutoff       = NULL,
                   M2CutoffQualityScore   = 30,
                   M2SlidingWindowSize    = 15)}
}
