% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassQualityReport.R
\docType{class}
\name{qualityReport-class}
\alias{qualityReport-class}
\title{qualityReport}
\description{
An S4 class for quality report for a SangeranalyseSeq S4 object
}
\section{Slots}{

\describe{
\item{\code{forward.read}}{.}
}}

\author{
Kuan-Hao Chao
}
