% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsQualityReport.R
\docType{methods}
\name{QualityReport-class-preQualityBasePlot}
\alias{QualityReport-class-preQualityBasePlot}
\alias{preQualityBasePlot,QualityReport-method}
\alias{QualityReport-class-qualityBasePlot}
\alias{qualityBasePlot,QualityReport-method}
\alias{QualityReport-class-updateQualityParam}
\alias{updateQualityParam,QualityReport-method}
\title{preQualityBasePlot}
\usage{
\S4method{preQualityBasePlot}{QualityReport}(object,
  readFeature = "Forward Read")

\S4method{qualityBasePlot}{QualityReport}(object)

\S4method{updateQualityParam}{QualityReport}(object,
  TrimmingMethod = "M1", M1TrimmingCutoff = 1e-04,
  M2CutoffQualityScore = NULL, M2SlidingWindowSize = NULL)
}
\arguments{
\item{object}{.}

\item{TrimmingMethod}{.}

\item{M1TrimmingCutoff}{.}

\item{M2CutoffQualityScore}{.}

\item{M2SlidingWindowSize}{.}

\item{object}{.}
}
\description{
A QualityReport method which pre-plots quality base interactive plot.

A QualityReport method which creates quality base interactive plot.

A QualityReport method which updates quality base interactive plot.
}
\examples{
\dontrun{preQualityBasePlot(qualityReport)}
\dontrun{qualityBasePlot(qualityReport)}
\dontrun{updateQualityParam(qualityReport,
                   TrimmingMethod         = "M2",
                   M1TrimmingCutoff       = NULL,
                   M2CutoffQualityScore   = 30,
                   M2SlidingWindowSize    = 15)}
}
