% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{readTable}
\alias{readTable}
\title{Method readTable}
\usage{
readTable(object, indentation = 0, ...)
}
\arguments{
\item{object}{A SangerRead, SangerContig, or SangerAlignment S4 instance.}

\item{indentation}{The indentation for different level printing}

\item{...}{Further generateReportSR-related parameters.}
}
\value{
None.
}
\description{
Method readTable
}
\examples{
data(sangerReadFData)
data(sangerContigData)
data(sangerAlignmentData)
\donttest{
readTable(sangerReadFData)
readTable(sangerContigData)
readTable(sangerAlignmentData)
}
}
