% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassAlignedConsensusSet.R
\docType{class}
\name{SangerAlignedConsensusSet-class}
\alias{SangerAlignedConsensusSet-class}
\title{SangerAlignedConsensusSet}
\description{
An S4 class for storing multiple single consensus reads to
}
\section{Slots}{

\describe{
\item{\code{parentDirectory}}{.}

\item{\code{suffixForwardRegExp}}{.}

\item{\code{suffixReverseRegExp}}{.}

\item{\code{consensusReadsList}}{.}

\item{\code{SCconsensusRead}}{.}

\item{\code{SCalignment}}{.}

\item{\code{SCdifferencesDF}}{.}

\item{\code{SCdistanceMatrix}}{.}

\item{\code{SCdendrogram}}{.}

\item{\code{SCindelsDF}}{.}

\item{\code{SCstopCodonsDF}}{.}

\item{\code{SCsecondaryPeakDF}}{.}
}}

\examples{
rawDataDir <- system.file("extdata", package = "sangeranalyseR")
inputFilesParentDir <- file.path(rawDataDir, "Allolobophora_chlorotica")
suffixForwardRegExp <- "_[F]_[0-9]*.ab1"
suffixReverseRegExp <- "_[R]_[0-9]*.ab1"
SangerAlignedConsensusSet <- new("SangerAlignedConsensusSet",
                     parentDirectory       = inputFilesParentDir,
                     suffixForwardRegExp   = suffixForwardRegExp,
                     suffixReverseRegExp   = suffixReverseRegExp,
                     TrimmingMethod        = "M2",
                     M1TrimmingCutoff      = NULL,
                     M2CutoffQualityScore  = 40,
                     M2SlidingWindowSize   = 10,
                     baseNumPerRow         = 100,
                     heightPerRow          = 200,
                     signalRatioCutoff     = 0.33,
                     showTrimmed           = TRUE)
}
\author{
Kuan-Hao Chao
}
