% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerAlignment.R
\docType{methods}
\name{SangerAlignment-class-launchAppSA}
\alias{SangerAlignment-class-launchAppSA}
\alias{launchAppSA,SangerAlignment-method}
\title{launchAppSA}
\usage{
\S4method{launchAppSA}{SangerAlignment}(object, outputDir = NULL)
}
\description{
A SangerAlignment method which launches Shiny app for SangerAlignment instance.
}
\examples{
\dontrun{RShinySA <- launchAppSA(sangerAlignment)}
}
