% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerAlignment.R
\docType{methods}
\name{SangerAlignment-class-writeFastaSA}
\alias{SangerAlignment-class-writeFastaSA}
\alias{writeFastaSA,SangerAlignment-method}
\title{writeFastaSA}
\usage{
\S4method{writeFastaSA}{SangerAlignment}(object, outputDir = NULL,
  compress = FALSE, compression_level = NA, selection = "all")
}
\arguments{
\item{object}{object}

\item{outputDir}{outputDir}

\item{compress}{compress}

\item{compression_level}{compression_level}

\item{selection}{selection}
}
\description{
A SangerAlignment method which writes sequences into Fasta files.
}
\examples{
\dontrun{writeFastaSA(sangerAlignment, "/Users/chaokuan-hao/Desktop/sangeranalyseR_fasta/SangerAlignment")}
}
