% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerAlignment.R
\docType{methods}
\name{SangerAlignment-class-updateQualityParam}
\alias{SangerAlignment-class-updateQualityParam}
\alias{SangerAlignment-class-launchAppSA}
\alias{SangerAlignment-class-writeFastaSA}
\alias{SangerAlignment-class-generateReportSA}
\title{updateQualityParam}
\usage{
\S4method{updateQualityParam}{SangerAlignment}(object,
  TrimmingMethod = "M1", M1TrimmingCutoff = 1e-04,
  M2CutoffQualityScore = NULL, M2SlidingWindowSize = NULL)

\S4method{launchAppSA}{SangerAlignment}(obj, outputDir = NULL)

\S4method{writeFastaSA}{SangerAlignment}(obj, outputDir = NULL,
  compress = FALSE, compression_level = NA, selection = "all")

\S4method{generateReportSA}{SangerAlignment}(obj, outputDir = NULL,
  includeSangerContig = TRUE, includeSangerRead = TRUE)
}
\description{
A SangerAlignment method which updates QualityReport parameter for each the SangerRead instance inside SangerAlignment.

A SangerAlignment method which launches Shiny app for SangerAlignment instance.

A SangerAlignment method which writes sequences into Fasta files.

A SangerAlignment method which generates final reports of the SangerContig instance.
}
\examples{
\dontrun{load("data/sangerAlignment.RData")
updateQualityParam(sangerAlignment,
                   TrimmingMethod         = "M2",
                   M1TrimmingCutoff       = NULL,
                   M2CutoffQualityScore   = 40,
                   M2SlidingWindowSize    = 15)}
\dontrun{load("data/sangerAlignment.RData")
RShinySA <- launchAppSA(sangerAlignment)}
\dontrun{load("data/sangerAlignment.RData")
writeFastaSA(sangerAlignment, "/Users/chaokuan-hao/Desktop/sangeranalyseR_fasta/SangerAlignment")}
\dontrun{load("data/sangerAlignment.RData")
generateReportSA(sangerAlignment)}
}
