% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangeranalyseSeq.R
\docType{class}
\name{SangeranalyseSeq-class}
\alias{SangeranalyseSeq-class}
\alias{SangeranalyseSeq}
\title{SangeranalyseSeq}
\description{
An S4 class extending sangerseq S4 class
}
\section{Slots}{

\describe{
\item{\code{qualityReport}}{.}
}}

\author{
Kuan-Hao Chao
}
