% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangerConsensusRead.R
\docType{class}
\name{SangerConsensusRead-class}
\alias{SangerConsensusRead-class}
\title{SangerConsensusRead}
\description{
An S4 class for storing multiple single reads to build up new
 consensus read
}
\section{Slots}{

\describe{
\item{\code{parentDirectory}}{.}

\item{\code{readsRegularExp}}{.}

\item{\code{cutoffQualityScore}}{.}

\item{\code{slidingWindowSize}}{.}
}}

\examples{
inputFilesParentDir <- system.file("extdata/", package = "sangeranalyseR")
samplesRegExp <- "ACHL"
A_chloroticConsensusReads <- new("SangerConsensusRead",
                                 parentDirectory = inputFilesParentDir,
                                 readsRegularExp = samplesRegExp,
                                 cutoffQualityScore  = 50L,
                                 slidingWindowSize   = 8L)
}
\author{
Kuan-Hao Chao
}
