% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerContig.R
\docType{methods}
\name{SangerContig-class-writeFastaSC}
\alias{SangerContig-class-writeFastaSC}
\title{writeFastaSC}
\usage{
\S4method{writeFastaSC}{SangerContig}(obj, outputDir = NULL,
  compress = FALSE, compression_level = NA, selection = "all")
}
\description{
writeFastaSC
}
\examples{
load("data/sangerContig.RData")
writeFastaSC(sangerContig)
}
