% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerContig.R
\docType{methods}
\name{SangerContig-class-updateQualityParam}
\alias{SangerContig-class-updateQualityParam}
\alias{SangerContig-class-launchAppSC}
\alias{SangerContig-class-writeFastaSC}
\alias{SangerContig-class-generateReportSC}
\title{updateQualityParam}
\usage{
\S4method{updateQualityParam}{SangerContig}(object,
  TrimmingMethod = "M1", M1TrimmingCutoff = 1e-04,
  M2CutoffQualityScore = NULL, M2SlidingWindowSize = NULL)

\S4method{launchAppSC}{SangerContig}(obj, outputDir = NULL)

\S4method{writeFastaSC}{SangerContig}(obj, outputDir = NULL,
  compress = FALSE, compression_level = NA, selection = "all")

\S4method{generateReportSC}{SangerContig}(obj, outputDir,
  includeSangerRead = TRUE, navigationAlignmentFN = NULL)
}
\description{
A SangerContig method which updates QualityReport parameter for each the SangerRead instance inside SangerContig.

A SangerContig method which launches Shiny app for SangerContig instance.

A SangerContig method which writes sequences into Fasta files.

A SangerContig method which generates final reports of the SangerContig instance.
}
\examples{
\dontrun{load("data/sangerContig.RData")
updateQualityParam(sangerContig,
                   TrimmingMethod         = "M2",
                   M1TrimmingCutoff       = NULL,
                   M2CutoffQualityScore   = 40,
                   M2SlidingWindowSize    = 15)}
\dontrun{load("data/sangerContig.RData")
RShinySC <- launchAppSC(sangerContig)}
\dontrun{load("data/sangerContig.RData")
writeFastaSC(sangerContig, "/Users/chaokuan-hao/Desktop/sangeranalyseR_fasta/SangerContig")}
\dontrun{load("data/sangerContig.RData")
generateReportSC(sangerContig)}
}
