% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangerMergeReads.R
\docType{class}
\name{SangerMergeReads-class}
\alias{SangerMergeReads-class}
\title{SangerMergeReads}
\description{
An S4 class for storing reads of sanger sequencing.
}
\section{Slots}{

\describe{
\item{\code{forwardReadSangerseq}}{.}

\item{\code{reverseReadSangerseq}}{.}
}}

\examples{
inputFilesPath <- system.file("extdata/", package = "sangeranalyseR")
A_chloroticaFdReadFN <- file.path(inputFilesPath,
                                  "Allolobophora_chlorotica",
                                  "ACHLO006-09[LCO1490_t1,HCO2198_t1]_F.ab1")
A_chloroticaRvReadFN <- file.path(inputFilesPath,
                                  "Allolobophora_chlorotica",
                                  "ACHLO006-09[LCO1490_t1,HCO2198_t1]_R.ab1")
A_chloroticaRead <- new("SangerMergeReads",
                        forwardReadFileName = A_chloroticaFdReadFN,
                        reverseReadFileName = A_chloroticaRvReadFN,
                        cutoffQualityScore  = 20L,
                        slidingWindowSize   = 8L)
}
\author{
Kuan-Hao Chao
}
