% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangerProject.R
\docType{class}
\name{SangerProject-class}
\alias{SangerProject-class}
\title{SangerProject}
\description{
An S4 class for storing multiple single consensus reads to
}
\section{Slots}{

\describe{
\item{\code{consensusReadsList}}{}
}}

\examples{
rawDataDir <- system.file("extdata", package = "sangeranalyseR")
inputFilesParentDir <- file.path(rawDataDir, "Allolobophora_chlorotica")
forwardRegExp <- "^ACHLO([0-9]*)-09\\\\[LCO1490_t1,HCO2198_t1\\\\]_F.ab1$"
reverseRegExp <- "^ACHLO([0-9]*)-09\\\\[LCO1490_t1,HCO2198_t1\\\\]_R.ab1$"
}
\author{
Kuan-Hao Chao
}
