% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\docType{methods}
\name{SangerRead-class-MakeBaseCalls}
\alias{SangerRead-class-MakeBaseCalls}
\alias{MakeBaseCalls,SangerRead-method}
\title{MakeBaseCalls}
\usage{
\S4method{MakeBaseCalls}{SangerRead}(object, signalRatioCutoff = 0.33)
}
\arguments{
\item{object}{object}

\item{signalRatioCutoff}{signalRatioCutoff}
}
\description{
A SangerRead method which does base calling on SangerRead instance
}
\examples{
data("sangerReadF")
MakeBaseCalls(sangerReadF, signalRatioCutoff = 0.22)
}
