% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\docType{methods}
\name{SangerRead-class-writeFastaSR}
\alias{SangerRead-class-writeFastaSR}
\alias{writeFastaSR,SangerRead-method}
\title{writeFastaSR}
\usage{
\S4method{writeFastaSR}{SangerRead}(object, outputDir = NULL,
  compress = FALSE, compression_level = NA)
}
\description{
A SangerRead method which writes the sequence into Fasta files.
}
\examples{
\dontrun{writeFastaSR(sangerReadF, "/Users/chaokuan-hao/Desktop/sangeranalyseR_fasta/SangerRead")}
}
