% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\docType{methods}
\name{SangerRead-class-qualityBasePlot}
\alias{SangerRead-class-qualityBasePlot}
\alias{SangerRead-class-updateQualityParam}
\alias{SangerRead-class-MakeBaseCalls}
\alias{SangerRead-class-writeFastaSR}
\alias{SangerRead-class-generateReportSR}
\title{qualityBasePlot}
\usage{
\S4method{qualityBasePlot}{SangerRead}(object)

\S4method{updateQualityParam}{SangerRead}(object, TrimmingMethod = "M1",
  M1TrimmingCutoff = 1e-04, M2CutoffQualityScore = NULL,
  M2SlidingWindowSize = NULL)

\S4method{MakeBaseCalls}{SangerRead}(obj, signalRatioCutoff = 0.33)

\S4method{writeFastaSR}{SangerRead}(obj, outputDir = NULL,
  compress = FALSE, compression_level = NA)

\S4method{generateReportSR}{SangerRead}(obj, outputDir,
  navigationContigFN = NULL, navigationAlignmentFN = NULL)
}
\description{
A SangerRead method which creates quality base interactive plot.

A SangerRead method which updates QualityReport parameter inside the SangerRead.

A SangerRead method which does base calling on SangerRead instance

A SangerRead method which writes the sequence into Fasta files.

A SangerRead method which generates final reports of the SangerRead instance.
}
\examples{
\dontrun{load("data/sangerRead.RData")
qualityBasePlot(sangerReadF)}
\dontrun{load("data/sangerRead.RData")
updateQualityParam(sangerReadF,
                   TrimmingMethod         = "M2",
                   M1TrimmingCutoff       = NULL,
                   M2CutoffQualityScore   = 40,
                   M2SlidingWindowSize    = 15)}
\dontrun{load("data/sangerRead.RData")
MakeBaseCalls(sangerReadF, signalRatioCutoff = 0.22)}
\dontrun{load("data/sangerRead.RData")
writeFastaSR(sangerReadF, "/Users/chaokuan-hao/Desktop/sangeranalyseR_fasta/SangerRead")}
\dontrun{load("data/sangerRead.RData")
generateReportSR(sangerReadF)}
}
