% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangerReads.R
\docType{class}
\name{SangerReads-class}
\alias{SangerReads-class}
\title{SangerReads}
\description{
An S4 class for storing reads of sanger sequencing.
}
\section{Slots}{

\describe{
\item{\code{forwardReadFileName}}{.}

\item{\code{reverseReadFileName}}{.}

\item{\code{forwardReadRawAbif}}{.}

\item{\code{reverseReadRawAbif}}{.}

\item{\code{forwardReadSangerseq}}{.}

\item{\code{reverseReadSangerseq}}{.}
}}

\examples{
inputFilesPath <- system.file("extdata/", package = "sangeranalyseR")
A_chloroticaFdReadFN <- file.path(inputFilesPath,
                                  "Allolobophora_chlorotica",
                                  "ACHLO006-09[LCO1490_t1,HCO2198_t1]_F.ab1")
A_chloroticaRvReadFN <- file.path(inputFilesPath,
                                  "Allolobophora_chlorotica",
                                  "ACHLO006-09[LCO1490_t1,HCO2198_t1]_R.ab1")
A_chloroticaRead <- new("SangerReads",
                        forwardReadFileName = A_chloroticaFdReadFN,
                        reverseReadFileName = A_chloroticaRvReadFN)
}
\author{
Kuan-Hao Chao
}
