% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangerSingleRead.R
\docType{class}
\name{sangerSingleRead-class}
\alias{sangerSingleRead-class}
\alias{sangerSingleRead}
\title{sangerSingleRead}
\description{
An S4 class extending sangerseq S4 class
}
\section{Slots}{

\describe{
\item{\code{qualityReport}}{.}
}}

\author{
Kuan-Hao Chao
}
