% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary.peaks.R
\name{secondary.peaks}
\alias{secondary.peaks}
\title{Check for secondary peaks in a sangerseq object}
\usage{
secondary.peaks(s, ratio = 0.33, output.folder = NA, file.prefix = "seq",
  processors = NULL)
}
\arguments{
\item{s}{a sangerseq s4 object from the sangerseqR package}

\item{ratio}{the ratio of the height of a secondary peak to a primary peak. Secondary peaks higher than this ratio are annotated. Those below the ratio are not.}

\item{output.folder}{If output.folder is NA (the default) no files are written. If a valid folder is provided, two files are written to that folder: a .csv file of the secondary peaks (see description below) and a .pdf file of the chromatogram.}

\item{file.prefix}{If output.folder is specified, this is the prefix which will be appended to the .csv and the .pdf file. The default is "seq".}

\item{processors}{The number of processors to use, or NULL (the default) for all available processors}
}
\value{
a data frame with one row per secondary peak above the ratio, and three columns: "position" is the position of the secondary peak relative to the primary sequence; "primary.basecall" is the primary base call; "secondary.basecall" is the secondary basecall.
}
\description{
This function finds and reports secondary peaks in a sangerseq object. It returns a table of secondary peaks, and optionally saves an annotated chromatogram and a csv file of the peak locations.
}
\keyword{chromatogram,}
\keyword{mismatch}
\keyword{peak,}

