% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ClassChromatogramParam.R,
%   R/ClassQualityReport.R
\docType{class}
\name{numericORNULL}
\alias{numericORNULL}
\alias{abifORNULL}
\alias{characterORNULL}
\alias{ChromatogramParamORNULL}
\alias{QualityReportORNULL}
\title{S4 Class Union slot types}
\description{
ClassUnion numericORNULL

ClassUnion abifORNULL

ClassUnion characterORNULL

ClassUnion ChromatogramParamORNULL

ClassUnion QualityReportORNULL
}
