% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.abi.file.R
\name{summarise.abi.file}
\alias{summarise.abi.file}
\title{Create a detailed summary of a single ABI sequencing file}
\usage{
summarise.abi.file(inputfile, trim.cutoff = 1e-04,
  secondary.peak.cutoff = 0.33, write.secondary.peak.files = FALSE,
  processors = NULL)
}
\description{
Create a detailed summary of a single ABI sequencing file
}
\keyword{chromatogram,}
\keyword{mismatch}
\keyword{peak,}

