% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.abi.folder.R
\name{summarise.abi.folder}
\alias{summarise.abi.folder}
\title{Create detailed summaries of all ABI sequencing reads in a folder (recursively searched)}
\usage{
summarise.abi.folder(input.folder, trim.cutoff = 1e-04,
  secondary.peak.ratio = 0.33, write.secondary.peak.files = FALSE,
  processors = NULL)
}
\description{
Create detailed summaries of all ABI sequencing reads in a folder (recursively searched)
}
\keyword{chromatogram,}
\keyword{mismatch}
\keyword{peak,}

