% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.merged.read.R
\name{summarise.merged.read}
\alias{summarise.merged.read}
\title{Produce a one-row summary of a merged read}
\usage{
summarise.merged.read(merged.read)
}
\arguments{
\item{merged.read}{a merged.read object produced by the merge.reads() function}
}
\value{
A numeric vector including (note that insertions, deletions, and stop codons will only be called if the read was merged with a reference amino acid sequence):
         \enumerate{
             \item {consensus.len}: the length of the consensus sequence\cr
             \item {n.reads}: the number of reads used to build the consensus\cr
             \item {distance.min}: the minimum distance between two reads \cr
             \item {distance.max}: the maximum distance between two reads \cr
             \item {distance.med}: the median distance between pairs of reads \cr
             \item {readlen.min}: the length of the smallest read \cr
             \item {readlen.max}: the length of the longest read \cr
             \item {readlen.med}: the median of all read lengths \cr
             \item {insertions.min}: the smallest number of insertions inferred in a read \cr
             \item {insertions.max}: the larges number of insertions inferred in a read \cr
             \item {insertions.med}: the median number of insertions inferred in a read \cr
             \item {deletions.min}: the smallest number of deletions inferred in a read \cr
             \item {deletions.max}: the larges number of deletions inferred in a read \cr
             \item {deletions.med}: the median number of deletions inferred in a read \cr
             \item {stops.min}: the smallest number of stop codons inferred in a read \cr
             \item {stops.max}: the larges number of stop codons inferred in a read \cr
             \item {stops.med}: the median number of stop codons inferred in a read \cr
         }
}
\description{
Produce a one-row summary of a merged read
}
