% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trim.mott.R
\name{trim.mott}
\alias{trim.mott}
\title{Trim a sequences with Mott's modified trimming algorithm}
\usage{
trim.mott(abif.seq, cutoff = 0.05, segment = 20)
}
\arguments{
\item{abif.seq}{a abif.seq s4 object from the sangerseqR package}

\item{cutoff}{default cutoff value for calculating base scores}

\item{segment}{minimum sequence length to return. If your input sequence is shorter than this, you just get your input sequence back.}
}
\value{
a vector of two integers: "trim_start" the start position of the sequence to keep; "trim_end" the end position of the sequence to keep
}
\description{
Removes low quality bases from the start and end of a sequence. This versino in R was ported from the Biopython implementation of the same algorithm: http://biopython.org/DIST/docs/api/Bio.SeqIO.AbiIO-module.html#_abi_trim. But note that the BioPython implementation always trims off the first base, while this implementation does not. For more information on the alogrithm, see http://www.phrap.org/phredphrap/phred.html http://www.clcbio.com/manual/genomics/Quality_abif_trimming.html
}
\keyword{phred,}
\keyword{quality,}
\keyword{trimming}

