% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.to.reference.R
\name{trim.to.reference}
\alias{trim.to.reference}
\title{Trim an aligned XStringSet object to the limits of a supplied reference sequence}
\usage{
trim.to.reference(stringset, reference)
}
\arguments{
\item{stringset}{an XStringSet object}

\item{reference}{an XStringSet object of the same type as the XStringSet object}
}
\description{
This function profile aligns the XStringSet object to the
reference sequence. It then trims the alignment to the limits of the 
reference sequence, and returns a trimmed alignment
}

