% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{writeFastaSA}
\alias{writeFastaSA}
\title{Method writeFastaSA}
\usage{
writeFastaSA(object, outputDir = NULL, compress = FALSE,
  compression_level = NA, selection = "all")
}
\arguments{
\item{object}{object}

\item{outputDir}{outputDir}

\item{compress}{compress}

\item{compression_level}{compression_level}

\item{selection}{selection}
}
\description{
Method writeFastaSA
}
