% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{writeFastaSR}
\alias{writeFastaSR}
\title{Method writeFastaSR}
\usage{
writeFastaSR(object, outputDir = NULL, compress = FALSE,
  compression_level = NA)
}
\arguments{
\item{object}{object}

\item{outputDir}{outputDir}

\item{compress}{compress}

\item{compression_level}{compression_level}
}
\description{
Method writeFastaSR
}
\examples{
data(sangerReadF)
writeFastaSR(sangerReadF)
}
