% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassChromatogramParam.R
\docType{class}
\name{ChromatogramParam-class}
\alias{ChromatogramParam-class}
\title{ChromatogramParam}
\description{
An S4 class for chromatogram parameter
}
\section{Slots}{

\describe{
\item{\code{baseNumPerRow}}{.}

\item{\code{signalRatioCutoff}}{.}

\item{\code{showTrimmed}}{.}
}}

\examples{
Chromatogram <- new("ChromatogramParam",
                     baseNumPerRow      = 100,
                     signalRatioCutoff  = 0.33,
                     showTrimmed        = TRUE)
}
\author{
Kuan-Hao Chao
}
