% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodShared.R
\name{generateReport}
\alias{generateReport}
\title{Method generateReport}
\usage{
generateReport(object, outputDir = NULL, includeSangerContig = TRUE,
  includeSangerRead = TRUE, ...)
}
\arguments{
\item{object}{object}

\item{outputDir}{outputDir}

\item{includeSangerContig}{includeSangerContig}

\item{includeSangerRead}{includeSangerRead}

\item{...}{...}
}
\description{
A method which generates final reports of the SangerRead,
 SangerContig, and SangerAlignment instance.
}
\examples{
data(sangerReadF)
data(sangerContig)
data(sangerAlignment)
\dontrun{
generateReport(sangerReadF)
generateReport(sangerContig)
generateReport(sangerAlignment)}
}
\author{
Kuan-Hao Chao
}
