% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodShared.R
\name{generateReport}
\alias{generateReport}
\title{Method generateReport}
\usage{
generateReport(object, outputDir = NULL, includeSangerContig = TRUE,
  includeSangerRead = TRUE, ...)
}
\arguments{
\item{object}{A SangerRead, SangerContig, or SangerAlignment S4 instance.}

\item{outputDir}{The output directory of the generated HTML report.}

\item{includeSangerContig}{The parameter that decides whether to include SangerContig level report. The value is \code{TRUE} or \code{FALSE} and the default is \code{TRUE}.}

\item{includeSangerRead}{The parameter that decides whether to include SangerRead level report. The value is \code{TRUE} or \code{FALSE} and the default is \code{TRUE}.}

\item{...}{Further generateReportSR, generateReportSC, and generateReportSA related parameters.}
}
\description{
A method which generates final reports of the SangerRead,
 SangerContig, and SangerAlignment instance.
}
\examples{
data(sangerReadFData)
data(sangerContigData)
data(sangerAlignmentData)
\dontrun{
generateReport(sangerReadFData)
generateReport(sangerContigData)
generateReport(sangerAlignmentData)}
}
\author{
Kuan-Hao Chao
}
