% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{generateReportSA}
\alias{generateReportSA}
\title{Method generateReportSA}
\usage{
generateReportSA(object, outputDir = NULL, includeSangerContig = TRUE,
  includeSangerRead = TRUE, ...)
}
\arguments{
\item{object}{A SangerAlignment S4 instance.}

\item{outputDir}{The output directory of the generated HTML report.}

\item{includeSangerContig}{The parameter that decides whether to include SangerContig level report. The value is \code{TRUE} or \code{FALSE} and the default is \code{TRUE}.}

\item{includeSangerRead}{The parameter that decides whether to include SangerRead level report. The value is \code{TRUE} or \code{FALSE} and the default is \code{TRUE}.}

\item{...}{Further generateReportSA-related parameters.}
}
\value{
The output absolute path to the SangerAlignment's HTML file.
}
\description{
Method generateReportSA
}
\examples{
data(sangerAlignmentData)
\dontrun{
generateReportSA(sangerAlignmentData)}
}
