% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{generateReportSA}
\alias{generateReportSA}
\title{Method generateReportSA}
\usage{
generateReportSA(object, outputDir = NULL, includeSangerContig = TRUE,
  includeSangerRead = TRUE)
}
\description{
Method generateReportSA
}
