% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{generateReportSC}
\alias{generateReportSC}
\title{Method generateReportSC}
\usage{
generateReportSC(object, outputDir = NULL, includeSangerRead = TRUE,
  ...)
}
\arguments{
\item{object}{A SangerContig S4 instance.}

\item{outputDir}{The output directory of the generated HTML report.}

\item{includeSangerRead}{The parameter that decides whether to include SangerRead level report. The value is \code{TRUE} or \code{FALSE} and the default is \code{TRUE}.}

\item{...}{Further generateReportSC-related parameters.}
}
\value{
The output absolute path to the SangerContig's HTML file.
}
\description{
Method generateReportSC
}
\examples{
data(sangerContigData)
\dontrun{
generateReportSC(sangerContigData)}
}
