% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{generateReportSC}
\alias{generateReportSC}
\title{Method generateReportSC}
\usage{
generateReportSC(object, outputDir = NULL, includeSangerRead = TRUE,
  ...)
}
\arguments{
\item{object}{object}

\item{outputDir}{outputDir}

\item{includeSangerRead}{includeSangerRead}

\item{...}{...}
}
\description{
Method generateReportSC
}
\examples{
data(sangerContig)
\dontrun{
generateReportSC(sangerContig)}
}
