% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{generateReportSR}
\alias{generateReportSR}
\title{Method generateReportSR}
\usage{
generateReportSR(object, outputDir = NULL, ...)
}
\arguments{
\item{object}{object}

\item{outputDir}{outputDir}

\item{...}{...}
}
\description{
Method generateReportSR
}
\examples{
data(sangerReadF)
\dontrun{
generateReportSR(sangerReadF)}
}
