% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{generateReportSR}
\alias{generateReportSR}
\title{Method generateReportSR}
\usage{
generateReportSR(object, outputDir = NULL, ...)
}
\arguments{
\item{object}{A SangerRead S4 instance.}

\item{outputDir}{The output directory of the generated HTML report.}

\item{...}{Further generateReportSR-related parameters.}
}
\description{
Method generateReportSR
}
\examples{
data(sangerReadFData)
\dontrun{
generateReportSR(sangerReadFData)}
}
