% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodShared.R
\name{launchApp}
\alias{launchApp}
\title{Method launchApp}
\usage{
launchApp(object, outputDir = NULL)
}
\arguments{
\item{object}{A SangerContig or SangerAlignment S4 instance.}

\item{outputDir}{The output directory of the saved new SangerContig or SangerAlignment S4 instance.}
}
\description{
A method which launches Shiny application of the SangerContig
 and SangerAlignment instance.
}
\examples{
data(sangerContigData)
data(sangerAlignmentData)
\dontrun{
launchApp(sangerContigData)
launchApp(sangerAlignmentData)}
}
\author{
Kuan-Hao Chao
}
