% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodShared.R
\name{launchApp}
\alias{launchApp}
\title{Method launchApp}
\usage{
launchApp(object, outputDir = NULL)
}
\arguments{
\item{object}{object}

\item{outputDir}{outputDir}
}
\description{
A method which launches Shiny application of the SangerContig
 and SangerAlignment instance.
}
\examples{
data(sangerContig)
data(sangerAlignment)
\dontrun{
launchApp(sangerContig)
launchApp(sangerAlignment)}
}
\author{
Kuan-Hao Chao
}
