% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{launchAppSA}
\alias{launchAppSA}
\title{Method launchAppSA}
\usage{
launchAppSA(object, outputDir = NULL)
}
\arguments{
\item{object}{A SangerAlignment S4 instance.}

\item{outputDir}{The output directory of the saved new SangerAlignment S4 instance.}
}
\value{
A \code{shiny.appobj} object.
}
\description{
Method launchAppSA
}
\examples{
data(sangerAlignmentData)
\dontrun{
launchAppSA(sangerAlignmentData)}
}
