% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{launchAppSC}
\alias{launchAppSC}
\title{Method launchAppSC}
\usage{
launchAppSC(object, outputDir = NULL)
}
\arguments{
\item{object}{object}

\item{outputDir}{outputDir}
}
\description{
Method launchAppSC
}
\examples{
data(sangerContig)
\dontrun{
launchAppSC(sangerContig)}
}
