% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{launchAppSC}
\alias{launchAppSC}
\title{Method launchAppSC}
\usage{
launchAppSC(object, outputDir = NULL)
}
\arguments{
\item{object}{A SangerContig S4 instance.}

\item{outputDir}{The output directory of the saved new SangerContig S4 instance.}
}
\description{
Method launchAppSC
}
\examples{
data(sangerContigData)
\dontrun{
launchAppSC(sangerContigData)}
}
