% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.readset.R
\name{make.readset}
\alias{make.readset}
\title{Make a readgroup object}
\usage{
make.readset(fwd.fnames, rev.fnames, trim = TRUE, trim.cutoff = 1e-04,
  max.secondary.peaks = NULL, secondary.peak.ratio = 0.33, min.length = 1,
  processors = NULL)
}
\arguments{
\item{fwd.fnames}{a list of full file paths to forward reads from ab1 files (i.e. those that do not need to be reverse-complemented).}

\item{rev.fnames}{a list of full file paths to reverse reads from ab1 files (i.e. those that *do* need to be reverse-complemented).}

\item{trim}{TRUE/FALSE trim sequences based on quality while creating readgroup? If TRUE, the trim.mott function is applied to each sequence before inclusion in the readgroup. Note, trimming only works if the raw data are stored in ab1 files with appropriate information.}

\item{trim.cutoff}{value passed to trim.mott as quality cutoff for sequencing trimming, only used if 'trim' == TRUE}

\item{max.secondary.peaks}{reads with more secondary peaks than this will not be included in the readset. The default (NULL) is to include all reads regardless of secondary peaks}

\item{secondary.peak.ratio}{Only applies if max.secondary.peaks is not NULL. The ratio of the height of a secondary peak to a primary peak. Secondary peaks higher than this ratio are counted. Those below the ratio are not.}

\item{min.length}{reads shorter than this will not be included in the readset. The default (1) means that all reads with length of 1 or more will be included.}

\item{processors}{The number of processors to use, or NULL (the default) for all available processors

A set of unaligned reads as a DNAstringset object, names are the input file paths.}
}
\description{
Make a readgroup object
}

