% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.reads.R
\name{make.readsets}
\alias{make.readsets}
\title{Automatically load readsets by grouping .ab1 files by name.}
\usage{
make.readsets(input.folder, forward.suffix, reverse.suffix, trim = TRUE,
  trim.cutoff = 1e-04, processors = NULL, min.length = 1,
  max.secondary.peaks = NULL, secondary.peak.ratio = 0.33)
}
\arguments{
\item{input.folder}{The parent folder of all of the reads contained in ab1 files you wish to analyse. Subfolders will be scanned recursively.}

\item{forward.suffix}{the suffix of the filenames for forward reads, i.e. reads that do not need to be reverse-complemented. Include the full suffix, e.g. "forward.ab1".}

\item{reverse.suffix}{the suffix of the filenames for reverse reads, i.e. reads that *do* need to be reverse-complemented. Include the full suffix, e.g. "reverse.ab1".}

\item{trim}{TRUE/FALSE trim sequences based on quality while creating readgroup? If TRUE, the trim.mott function is applied to each sequence before inclusion in the readgroup. Note, trimming only works if the raw data are stored in ab1 files with appropriate information.}

\item{trim.cutoff}{value passed to trim.mott as quality cutoff for sequencing trimming, only used if 'trim' == TRUE}

\item{processors}{The number of processors to use, or NULL (the default) for all available processors}

\item{min.length}{reads shorter than this will not be included in the readset. The default (1) means that all reads with length of 1 or more will be included.}

\item{max.secondary.peaks}{reads with more secondary peaks than this will not be included in the readset. The default (NULL) is to include all reads regardless of secondary peaks}

\item{secondary.peak.ratio}{Only applies if max.secondary.peaks is not NULL. The ratio of the height of a secondary peak to a primary peak. Secondary peaks higher than this ratio are counted. Those below the ratio are not.}
}
\description{
Automatically load readsets by grouping .ab1 files by name.
}

