% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{MakeBaseCalls}
\alias{MakeBaseCalls}
\title{Method MakeBaseCalls}
\usage{
MakeBaseCalls(object, signalRatioCutoff = 0.33)
}
\arguments{
\item{object}{A SangerRead S4 instance.}

\item{signalRatioCutoff}{The ratio of the height of a secondary peak to a primary peak. Secondary peaks higher than this ratio are annotated. Those below the ratio are excluded. The default value is \code{0.33}.}
}
\value{
A \code{SangerRead} instance.
}
\description{
Method MakeBaseCalls
}
\examples{
data(sangerReadFData)
MakeBaseCalls(sangerReadFData, signalRatioCutoff = 0.22)
}
