% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.reads.R
\name{merge.reads}
\alias{merge.reads}
\title{Merge fwd and reverse reads into a single sequence}
\usage{
\method{merge}{reads}(fwd.seqs, rev.seqs, ref.aa.seq = NULL,
  minInformation = 1, minReads = 0, processors = NULL,
  genetic.code = GENETIC_CODE)
}
\arguments{
\item{fwd.seqs}{a list of forward sequences, each should be a sangerseq s4 object from the sangerseqR package. These sequences will not be reverse-complemented before alignment.}

\item{rev.seqs}{a list of reverse sequences, each should be a sangerseq s4 object from the sangerseqR package. These sequences will be reverse-complemented before alignment.}

\item{ref.aa.seq}{an amino acid reference sequence supplied as a string or an AAString object. If your sequences are protein-coding DNA seuqences, and you want to have frameshifts automatically detected and corrected, supply a reference amino acid sequence via this argument. If this argument is supplied, the sequences are then kept in frame for the alignment step. Fwd sequences are assumed to come from the sense (i.e. coding, or "+") strand.}

\item{minInformation}{minimum fraction of the sequences required to call a consensus sequence at any given position (see the ConsensusSequence() function from DECIPHER for more information). Defaults to 1.0 for this function, implying that the consensus sequence is only called when all fwd and rev reads overlap.}

\item{minReads}{minimum number of reads that are required to call a consensus sequence at any given position. This can be used in place of minInformation if you set minInformation to 0.0. For example, if you expect to have non-overlapping reads making up your consensus, but only wish to call a consensus where at least two reads overlap, set minInformation to 0.0 and minReads to 2. If you set both minInformation and minReads, the value corresponding to the larger number of reads will be used.}

\item{processors}{The number of processors to use, or NULL (the default) for all available processors}

\item{genetic.code}{Named character vector in the same format as GENETIC_CODE (the default), which represents the standard genetic code. This is the code with which the function will attempt to translate your DNA sequences. You can get an appropriate vector with the getGeneticCode() function. The default is the standard code.}
}
\value{
a list with the following components:
         \enumerate{
             \item consensus: the consensus sequence from the merged reads.\cr
             \item alignment: the alignment of all the reads, with the called consensus sequence. E.g. if the list was called 'merged.reads', you could use BrowseSeqs(merged.reads$alignment) to view the alignment.
             \item differences: a data frame of the number of pairwise differences between each read and the consensus sequence, as well as the number of bases in each input read that did not contribute to the consensus sequence. Can assist in detecting incorrect reads, or reads with a lot of errors.
             \item distance.matrix: a distance matrix of genetic distances (corrected with the JC model) between all of the input reads.
             \item dendrogram: a dendrogram depicting the distance.matrix. E.g. if the list was called 'merged.reads', you could use plot(merged.reads$dendrogram) to see the dendrogram.
             \item indels: if you specified a reference sequence via ref.aa.seq, then this will be a data frame describing the number of indels and deletions that were made to each of the input reads in order to correct frameshift mutations.
         }
}
\description{
This function attempts to merge any number of forward and reverse sequences into a single consensus sequence. It calculates a number of statistics that should help you decide whether a sensible consensus sequence exists for your data.\cr
Note that sequence names can be provided via the names on the input lists. If this is done, the names will be preserved in the output data. Otherwise, sequences will be called: fwd_1, fwd_2, etc.
}
\keyword{alignment,}
\keyword{merge,}
\keyword{reads,}
\keyword{sequence}

