% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.reads.R
\name{merge.reads}
\alias{merge.reads}
\title{Merge sequence reads into a single consensus sequence}
\usage{
\method{merge}{reads}(readset, ref.aa.seq = NULL, minInformation = 0.75,
  threshold = 0.05, processors = NULL, genetic.code = GENETIC_CODE)
}
\arguments{
\item{readset}{a DNAStringSet object with at least 2 reads}

\item{ref.aa.seq}{an amino acid reference sequence supplied as a string or an AAString object. If your sequences are protein-coding DNA seuqences, and you want to have frameshifts automatically detected and corrected, supply a reference amino acid sequence via this argument. If this argument is supplied, the sequences are then kept in frame for the alignment step. Fwd sequences are assumed to come from the sense (i.e. coding, or "+") strand.}

\item{minInformation}{minimum fraction of the sequences required to call a consensus sequence at any given position (see the ConsensusSequence() function from DECIPHER for more information). Defaults to 0.75implying that 3/4 of all reads must be present in order to call a consensus.}

\item{threshold}{Numeric giving the maximum fraction of sequence information that can be lost in the consensus sequence (see the ConsensusSequence() function from DECIPHER for more information). Defaults to 0.05, implying that each consensus base can ignore at most 5 percent of the information at a given position.}

\item{processors}{The number of processors to use, or NULL (the default) for all available processors}

\item{genetic.code}{Named character vector in the same format as GENETIC_CODE (the default), which represents the standard genetic code. This is the code with which the function will attempt to translate your DNA sequences. You can get an appropriate vector with the getGeneticCode() function. The default is the standard code.}
}
\value{
A list with the following components:
         \enumerate{
             \item {consensus}: the consensus sequence from the merged reads.\cr
             \item {alignment}: the alignment of all the reads, with the called consensus sequence. E.g. if the list was called 'merged.reads', you could use BrowseSeqs(merged.reads$alignment) to view the alignment.
             \item {differences}: a data frame of the number of pairwise differences between each read and the consensus sequence, as well as the number of bases in each input read that did not contribute to the consensus sequence. Can assist in detecting incorrect reads, or reads with a lot of errors.
             \item {distance.matrix}: a distance matrix of genetic distances (corrected with the JC model) between all of the input reads.
             \item {dendrogram}: a dendrogram depicting the distance.matrix. E.g. if the list was called 'merged.reads', you could use plot(merged.reads$dendrogram) to see the dendrogram.
             \item {indels}: if you specified a reference sequence via ref.aa.seq, then this will be a data frame describing the number of indels and deletions that were made to each of the input reads in order to correct frameshift mutations.
         }
}
\description{
This function attempts to merge any number of unaligned reads into a single consensus sequence. It calculates a number of statistics that should help you decide whether a sensible consensus sequence exists for your data.\cr
}
\keyword{alignment,}
\keyword{merge,}
\keyword{reads,}
\keyword{sequence}

