% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassObjectResults.R
\docType{class}
\name{ObjectResults-class}
\alias{ObjectResults-class}
\title{ObjectResults}
\description{
An S4 class storing results related inputs in a SangerRead, SangerContig, and SangerAlignment S4 object.
}
\section{Slots}{

\describe{
\item{\code{printLevel}}{}
}}

\examples{
objectResults <- new("ObjectResults",
                     creationResult   = TRUE,
                     errorMessages    = character(0),
                     errorTypes       = character(0),
                     warningMessages  = character(0),
                     warningTypes     = character(0),
                     readResultTable =  data.frame(),
                     printLevel       = "SangerRead")
}
\author{
Kuan-Hao Chao
}
