% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MethodSangerRead.R
\name{primaryAASeqS2}
\alias{primaryAASeqS2}
\alias{primaryAASeqS2,SangerRead-method}
\title{Method primaryAASeqS2}
\usage{
primaryAASeqS2(object)

\S4method{primaryAASeqS2}{SangerRead}(object)
}
\arguments{
\item{object}{A SangerRead S4 instance.}
}
\value{
Frame-2 AA translation as \code{AAString} (lazy).
}
\description{
Method primaryAASeqS2
}
\examples{
data(sangerReadFData)
primaryAASeqS2(sangerReadFData)
}
