% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsQualityReport.R
\docType{methods}
\name{QualityReport-class-qualityBasePlot}
\alias{QualityReport-class-qualityBasePlot}
\alias{qualityBasePlot,QualityReport-method}
\title{qualityBasePlot}
\usage{
\S4method{qualityBasePlot}{QualityReport}(object)
}
\description{
A QualityReport method which creates quality base interactive plot.
}
\examples{
\dontrun{qualityBasePlot(qualityReport)}
}
