% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsQualityReport.R
\docType{methods}
\name{QualityReport-class-qualityBasePlot}
\alias{QualityReport-class-qualityBasePlot}
\title{qualityBasePlot}
\usage{
\S4method{qualityBasePlot}{QualityReport}(object)
}
\description{
qualityBasePlot
}
\examples{
\dontrun{
load("data/qualityReport.RData")
qualityBasePlot(qualityReport)
}
}
