% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsQualityReport.R
\docType{methods}
\name{QualityReport-class-updateQualityParam}
\alias{QualityReport-class-updateQualityParam}
\title{updateQualityParam}
\usage{
\S4method{updateQualityParam}{QualityReport}(object,
  TrimmingMethod = "M1", M1TrimmingCutoff = 1e-04,
  M2CutoffQualityScore = NULL, M2SlidingWindowSize = NULL)
}
\description{
updateQualityParam
}
\examples{
\dontrun{
load("data/sangerRead.RData")
updateQualityParam(qualityReport,
                   TrimmingMethod         = "M2",
                   M1TrimmingCutoff       = NULL,
                   M2CutoffQualityScore   = 30,
                   M2SlidingWindowSize    = 15)
}
}
