% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassQualityReport.R
\docType{class}
\name{QualityReport-class}
\alias{QualityReport-class}
\title{QualityReport}
\description{
An S4 class for quality report for a SangerSingleRead S4 object
}
\section{Slots}{

\describe{
\item{\code{readFeature}}{.}

\item{\code{qualityPhredScores}}{.}

\item{\code{qualityBaseScore}}{.}

\item{\code{trimmingStartPos}}{.}

\item{\code{trimmingFinishPos}}{.}

\item{\code{cutoffQualityScore}}{.}

\item{\code{slidingWindowSize}}{.}
}}

\examples{
inputFilesPath <- system.file("extdata/", package = "sangeranalyseR")
A_chloroticaFdReadFN <- file.path(inputFilesPath,
                                  "Allolobophora_chlorotica",
                                  "ACHLO006-09[LCO1490_t1,HCO2198_t1]_F.ab1")
A_chloroticaRead <-
       SangerSingleRead(readFeature         = "ForwardRead",
                        readFileName        = A_chloroticaFdReadFN,
                        cutoffQualityScore  = 60L,
                        slidingWindowSize   = 8L)
"@"(A_chloroticaRead, QualityReport)
}
\author{
Kuan-Hao Chao
}
