% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassQualityReport.R
\docType{class}
\name{QualityReport-class}
\alias{QualityReport-class}
\title{QualityReport}
\description{
An S4 class for quality report for a SangerSingleRead S4 object
}
\section{Slots}{

\describe{
\item{\code{readFeature}}{.}

\item{\code{qualityPhredScores}}{.}

\item{\code{qualityBaseScore}}{.}

\item{\code{rawSeqLength}}{.}

\item{\code{trimmedSeqLength}}{.}

\item{\code{trimmedStartPos}}{.}

\item{\code{trimmedFinishPos}}{.}

\item{\code{rawMeanQualityScore}}{.}

\item{\code{trimmedMeanQualityScore}}{.}

\item{\code{rawMinQualityScore}}{.}

\item{\code{trimmedMinQualityScore}}{.}

\item{\code{TrimmingMethod}}{.}

\item{\code{M1TrimmingCutoff}}{.}

\item{\code{M2CutoffQualityScore}}{.}

\item{\code{M2SlidingWindowSize}}{.}
}}

\examples{
inputFilesPath <- system.file("extdata/", package = "sangeranalyseR")
A_chloroticaFdReadFN <- file.path(inputFilesPath,
                                  "Allolobophora_chlorotica",
                                  "ACHLO006-09[LCO1490_t1,HCO2198_t1]_F_1.ab1")
A_chloroticaRead <-
       SangerSingleRead(readFeature         = "ForwardRead",
                        readFileName        = A_chloroticaFdReadFN,
                        M2CutoffQualityScore  = 60L,
                        M2SlidingWindowSize   = 8L)
"@"(A_chloroticaRead, QualityReport)
}
\author{
Kuan-Hao Chao
}
