% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassQualityReport.R
\docType{class}
\name{QualityReport-class}
\alias{QualityReport-class}
\title{QualityReport}
\description{
An S4 class for quality report for a SangerSingleRead S4 object
}
\section{Slots}{

\describe{
\item{\code{qualityScoresID}}{.}

\item{\code{qualityPhredScores}}{.}

\item{\code{qualityPhredScoresBC}}{.}

\item{\code{qualityBaseScores}}{.}

\item{\code{qualityBaseScoresBC}}{.}

\item{\code{rawSeqLength}}{.}

\item{\code{trimmedSeqLength}}{.}

\item{\code{trimmedStartPos}}{.}

\item{\code{trimmedFinishPos}}{.}

\item{\code{rawMeanQualityScore}}{.}

\item{\code{trimmedMeanQualityScore}}{.}

\item{\code{rawMinQualityScore}}{.}

\item{\code{trimmedMinQualityScore}}{.}

\item{\code{remainingRatio}}{.}

\item{\code{TrimmingMethod}}{.}

\item{\code{M1TrimmingCutoff}}{.}

\item{\code{M2CutoffQualityScore}}{.}

\item{\code{M2SlidingWindowSize}}{.}
}}

\examples{
inputFilesPath <- system.file("extdata/", package = "sangeranalyseR")
A_chloroticaFdReadFN <- file.path(inputFilesPath,
                             "Allolobophora_chlorotica",
                             "ACHLO006-09[LCO1490_t1,HCO2198_t1]_F_1.ab1")
A_chloroticaRead <-
       SangerSingleRead(readFeature           = "Forward Read",
                        readFileName          = A_chloroticaFdReadFN,
                        TrimmingMethod        = "M2",
                        M1TrimmingCutoff      = NULL,
                        M2CutoffQualityScore  = 40,
                        M2SlidingWindowSize   = 10)
"@"(A_chloroticaRead, QualityReport)
}
\author{
Kuan-Hao Chao
}
