% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerAlignment.R
\docType{methods}
\name{SangerAlignment-class-generateReportSA}
\alias{SangerAlignment-class-generateReportSA}
\alias{generateReportSA,SangerAlignment-method}
\title{generateReportSA}
\usage{
\S4method{generateReportSA}{SangerAlignment}(object, outputDir = NULL,
  includeSangerContig = TRUE, includeSangerRead = TRUE)
}
\description{
A SangerAlignment method which generates final reports of the SangerContig instance.
}
\examples{
\dontrun{generateReportSA(sangerAlignment)}
}
