% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerAlignment.R
\docType{methods}
\name{SangerAlignment-class-generateReportSA}
\alias{SangerAlignment-class-generateReportSA}
\alias{generateReportSA,SangerAlignment-method}
\title{generateReportSA}
\usage{
\S4method{generateReportSA}{SangerAlignment}(object, outputDir,
  includeSangerContig = TRUE, includeSangerRead = TRUE)
}
\arguments{
\item{object}{object}

\item{outputDir}{outputDir}

\item{includeSangerContig}{includeSangerContig}

\item{includeSangerRead}{includeSangerRead}
}
\description{
A SangerAlignment method which generates final reports of the SangerContig instance.
}
\examples{
data("sangerAlignment")
\dontrun{
generateReportSA(sangerAlignment)}
}
