% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerAlignment.R
\docType{methods}
\name{SangerAlignment-class-generateReportSA}
\alias{SangerAlignment-class-generateReportSA}
\alias{generateReportSA,SangerAlignment-method}
\title{generateReportSA}
\usage{
\S4method{generateReportSA}{SangerAlignment}(object, outputDir,
  includeSangerContig = TRUE, includeSangerRead = TRUE)
}
\arguments{
\item{object}{A SangerAlignment S4 instance.}

\item{outputDir}{The output directory of the generated HTML report.}

\item{includeSangerContig}{The parameter that decides whether to include SangerContig level report. The value is \code{TRUE} or \code{FALSE} and the default is \code{TRUE}.}

\item{includeSangerRead}{The parameter that decides whether to include SangerRead level report. The value is \code{TRUE} or \code{FALSE} and the default is \code{TRUE}.}
}
\description{
A SangerAlignment method which generates final reports of the SangerContig instance.
}
\examples{
data("sangerAlignmentData")
\dontrun{
generateReportSA(sangerAlignmentData)}
}
