% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerAlignment.R
\docType{methods}
\name{SangerAlignment-class-generateReportSA}
\alias{SangerAlignment-class-generateReportSA}
\title{generateReportSA}
\usage{
\S4method{generateReportSA}{SangerAlignment}(obj, outputDir = NULL,
  includeSangerContig = TRUE, includeSangerRead = TRUE)
}
\description{
generateReportSA
}
\examples{
\dontrun{
load("data/sangerAlignment.RData")
generateReportSA(sangerAlignment)
}
}
