% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerAlignment.R
\docType{methods}
\name{SangerAlignment-class-launchAppSA}
\alias{SangerAlignment-class-launchAppSA}
\title{launchAppSA}
\usage{
\S4method{launchAppSA}{SangerAlignment}(obj, outputDir = NULL)
}
\description{
launchAppSA
}
\examples{
load("data/sangerAlignment.RData")
RShinySA <- launchAppSA(sangerAlignment)
}
