% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerAlignment.R
\docType{methods}
\name{SangerAlignment-class-updateQualityParam}
\alias{SangerAlignment-class-updateQualityParam}
\title{updateQualityParam}
\usage{
\S4method{updateQualityParam}{SangerAlignment}(object,
  TrimmingMethod = "M1", M1TrimmingCutoff = 1e-04,
  M2CutoffQualityScore = NULL, M2SlidingWindowSize = NULL)
}
\description{
updateQualityParam
}
\examples{
\dontrun{load("data/sangerAlignment.RData")
updateQualityParam(sangerAlignment,
                   TrimmingMethod         = "M2",
                   M1TrimmingCutoff       = NULL,
                   M2CutoffQualityScore   = 40,
                   M2SlidingWindowSize    = 15)}
}
