% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerAlignment.R
\docType{methods}
\name{SangerAlignment-class-updateQualityParam}
\alias{SangerAlignment-class-updateQualityParam}
\alias{updateQualityParam,SangerAlignment-method}
\title{updateQualityParam}
\usage{
\S4method{updateQualityParam}{SangerAlignment}(object,
  TrimmingMethod = "M1", M1TrimmingCutoff = 1e-04,
  M2CutoffQualityScore = NULL, M2SlidingWindowSize = NULL)
}
\description{
A SangerAlignment method which updates QualityReport parameter for each the SangerRead instance inside SangerAlignment.
}
\examples{
\dontrun{updateQualityParam(sangerAlignment,
                   TrimmingMethod         = "M2",
                   M1TrimmingCutoff       = NULL,
                   M2CutoffQualityScore   = 40,
                   M2SlidingWindowSize    = 15)}
}
