% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerAlignment.R
\docType{methods}
\name{SangerAlignment-class-writeFastaSA}
\alias{SangerAlignment-class-writeFastaSA}
\title{writeFastaSA}
\usage{
\S4method{writeFastaSA}{SangerAlignment}(obj, outputDir = NULL,
  compress = FALSE, compression_level = NA, selection = "all")
}
\description{
writeFastaSA
}
\examples{
load("data/sangerAlignment.RData")
writeFastaSA(sangerAlignment)
}
