% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangerAlignment.R
\docType{class}
\name{SangerAlignment-class}
\alias{SangerAlignment-class}
\title{SangerAlignment}
\description{
An S4 class for storing multiple single consensus reads to
}
\section{Slots}{

\describe{
\item{\code{parentDirectory}}{.}

\item{\code{suffixReverseRegExp}}{.}

\item{\code{minFractionCallSCSet}}{.}

\item{\code{maxFractionLostSCSet}}{.}

\item{\code{geneticCode}}{.}

\item{\code{refAminoAcidSeq}}{.}

\item{\code{contigList}}{.}

\item{\code{contigsConsensus}}{.}

\item{\code{contigsAlignment}}{.}

\item{\code{contigsTree}}{.}
}}

\examples{
rawDataDir <- system.file("extdata", package = "sangeranalyseR")
inputFilesParentDir <- file.path(rawDataDir, "Allolobophora_chlorotica")
suffixForwardRegExp <- "_[F]_[0-9]*.ab1"
suffixReverseRegExp <- "_[R]_[0-9]*.ab1"
SangerAlignment <- new("SangerAlignment",
                               parentDirectory       = rawDataDir,
                               suffixForwardRegExp   = suffixForwardRegExp,
                               suffixReverseRegExp   = suffixReverseRegExp,
                               refAminoAcidSeq = "SRQWLFSTNHKDIGTLYFIFGAWAGMVGTSLSILIRAELGHPGALIGDDQIYNVIVTAHAFIMIFFMVMPIMIGGFGNWLVPLMLGAPDMAFPRMNNMSFWLLPPALSLLLVSSMVENGAGTGWTVYPPLSAGIAHGGASVDLAIFSLHLAGISSILGAVNFITTVINMRSTGISLDRMPLFVWSVVITALLLLLSLPVLAGAITMLLTDRNLNTSFFDPAGGGDPILYQHLFWFFGHPEVYILILPGFGMISHIISQESGKKETFGSLGMIYAMLAIGLLGFIVWAHHMFTVGMDVDTRAYFTSATMIIAVPTGIKIFSWLATLHGTQLSYSPAILWALGFVFLFTVGGLTGVVLANSSVDIILHDTYYVVAHFHYVLSMGAVFAIMAGFIHWYPLFTGLTLNNKWLKSHFIIMFIGVNLTFFPQHFLGLAGMPRRYSDYPDAYTTWNIVSTIGSTISLLGILFFFFIIWESLVSQRQVIYPIQLNSSIEWYQNTPPAEHSYSELPLLTN",
                               TrimmingMethod        = "M2",
                               M1TrimmingCutoff      = NULL,
                               M2CutoffQualityScore  = 40,
                               M2SlidingWindowSize   = 10,
                               baseNumPerRow         = 100,
                               heightPerRow          = 200,
                               signalRatioCutoff     = 0.33,
                               showTrimmed           = TRUE)
}
\author{
Kuan-Hao Chao
}
