% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sangerAlignmentData}
\alias{sangerAlignmentData}
\title{SangerAlignment instance}
\usage{
data(sangerAlignmentData)
}
\description{
SangerAlignment instance
}
\author{
Kuan-Hao Chao
}
\keyword{datasets}
