% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangerConsensusRead.R
\docType{class}
\name{SangerConsensusRead-class}
\alias{SangerConsensusRead-class}
\title{SangerConsensusRead}
\description{
An S4 class for storing multiple single reads to build up new
 consensus read
}
\section{Slots}{

\describe{
\item{\code{readingFrame}}{}
}}

\examples{
rawDataDir <- system.file("extdata", package = "sangeranalyseR")
inputFilesParentDir <- file.path(rawDataDir, "Allolobophora_chlorotica")
forwardRegExp <- "^ACHLO([0-9]*)-09\\\\[LCO1490_t1,HCO2198_t1\\\\]_F.ab1$"
reverseRegExp <- "^ACHLO([0-9]*)-09\\\\[LCO1490_t1,HCO2198_t1\\\\]_R.ab1$"
A_chloroticConsensusReads <- new("SangerConsensusRead",
                                 parentDirectory       = inputFilesParentDir,
                                 forwardReadsRegularExp= forwardRegExp,
                                 reverseReadsRegularExp= reverseRegExp,
                                 cutoffQualityScore    = 50,
                                 slidingWindowSize     = 8)
}
\author{
Kuan-Hao Chao
}
