% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangerConsensusRead.R
\docType{class}
\name{SangerConsensusRead-class}
\alias{SangerConsensusRead-class}
\title{SangerConsensusRead}
\description{
An S4 class for storing multiple single reads to build up new
 consensus read
}
\section{Slots}{

\describe{
\item{\code{consenesusReadName}}{.}

\item{\code{suffixForwardRegExp}}{.}

\item{\code{suffixReverseRegExp}}{.}

\item{\code{secondaryPeakDF}}{}
}}

\examples{
rawDataDir <- system.file("extdata", package = "sangeranalyseR")
inputFilesParentDir <- file.path(rawDataDir, "Allolobophora_chlorotica")
consenesusReadName <- "ACHLO006-09[LCO1490_t1,HCO2198_t1]"
suffixForwardRegExp <- "_[F]_[0-9]*.ab1"
suffixReverseRegExp <- "_[R]_[0-9]*.ab1"
A_chloroticConsensusReads <- new("SangerConsensusRead",
                                 parentDirectory       = inputFilesParentDir,
                                 consenesusReadName    = consenesusReadName,
                                 suffixForwardRegExp   = suffixForwardRegExp,
                                 suffixReverseRegExp   = suffixReverseRegExp,
                                 TrimmingMethod        = "M2",
                                 M1TrimmingCutoff      = NULL,
                                 M2CutoffQualityScore  = 40,
                                 M2SlidingWindowSize   = 10)
}
\author{
Kuan-Hao Chao
}
