% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerContig.R
\docType{methods}
\name{SangerContig-class-generateReportSC}
\alias{SangerContig-class-generateReportSC}
\alias{generateReportSC,SangerContig-method}
\title{generateReportSC}
\usage{
\S4method{generateReportSC}{SangerContig}(object, outputDir,
  includeSangerRead = TRUE, navigationAlignmentFN = NULL)
}
\arguments{
\item{object}{object}

\item{outputDir}{outputDir}

\item{includeSangerRead}{includeSangerRead}

\item{navigationAlignmentFN}{navigationAlignmentFN}
}
\description{
A SangerContig method which generates final reports of the SangerContig instance.
}
\examples{
\dontrun{generateReportSC(sangerContig)}
}
