% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerContig.R
\docType{methods}
\name{SangerContig-class-generateReportSC}
\alias{SangerContig-class-generateReportSC}
\title{generateReportSC}
\usage{
\S4method{generateReportSC}{SangerContig}(obj, outputDir,
  includeSangerRead = TRUE, navigationAlignmentFN = NULL)
}
\description{
generateReportSC
}
\examples{
\dontrun{
load("data/sangerContig.RData")
generateReportSC(sangerContig)
}
}
