% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerContig.R
\docType{methods}
\name{SangerContig-class-generateReportSC}
\alias{SangerContig-class-generateReportSC}
\alias{generateReportSC,SangerContig-method}
\title{generateReportSC}
\usage{
\S4method{generateReportSC}{SangerContig}(object, outputDir,
  includeSangerRead = TRUE, navigationAlignmentFN = NULL)
}
\arguments{
\item{object}{A SangerContig S4 instance.}

\item{outputDir}{The output directory of the generated HTML report.}

\item{includeSangerRead}{The parameter that decides whether to include SangerRead level report. The value is \code{TRUE} or \code{FALSE} and the default is \code{TRUE}.}

\item{navigationAlignmentFN}{The internal parameter passed to HTML report. Users should not modify this parameter on their own.}
}
\value{
The output absolute path to the SangerContig's HTML file.
}
\description{
A SangerContig method which generates final reports of the SangerContig instance.
}
\examples{
data("sangerContigData")
\dontrun{
generateReportSC(sangerContigData)}
}
