% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerContig.R
\docType{methods}
\name{SangerContig-class-launchAppSC}
\alias{SangerContig-class-launchAppSC}
\title{launchAppSC}
\usage{
\S4method{launchAppSC}{SangerContig}(obj, outputDir = NULL)
}
\description{
launchAppSC
}
\examples{
\dontrun{load("data/sangerContig.RData")
RShinySC <- launchAppSC(sangerContig)}
}
