% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerContig.R
\docType{methods}
\name{SangerContig-class-launchAppSC}
\alias{SangerContig-class-launchAppSC}
\alias{launchAppSC,SangerContig-method}
\title{launchAppSC}
\usage{
\S4method{launchAppSC}{SangerContig}(object, outputDir = NULL)
}
\description{
A SangerContig method which launches Shiny app for SangerContig instance.
}
\examples{
\dontrun{RShinySC <- launchAppSC(sangerContig)}
}
