% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerContig.R
\docType{methods}
\name{SangerContig-class-updateQualityParam}
\alias{SangerContig-class-updateQualityParam}
\alias{updateQualityParam,SangerContig-method}
\title{updateQualityParam}
\usage{
\S4method{updateQualityParam}{SangerContig}(object,
  TrimmingMethod = "M1", M1TrimmingCutoff = 1e-04,
  M2CutoffQualityScore = NULL, M2SlidingWindowSize = NULL)
}
\description{
A SangerContig method which updates QualityReport parameter for each the SangerRead instance inside SangerContig.
}
\examples{
\dontrun{updateQualityParam(sangerContig,
                   TrimmingMethod         = "M2",
                   M1TrimmingCutoff       = NULL,
                   M2CutoffQualityScore   = 40,
                   M2SlidingWindowSize    = 15)}
}
