% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerContig.R
\docType{methods}
\name{SangerContig-class-updateQualityParam}
\alias{SangerContig-class-updateQualityParam}
\alias{updateQualityParam,SangerContig-method}
\title{updateQualityParam}
\usage{
\S4method{updateQualityParam}{SangerContig}(object,
  TrimmingMethod = "M1", M1TrimmingCutoff = 1e-04,
  M2CutoffQualityScore = NULL, M2SlidingWindowSize = NULL)
}
\arguments{
\item{object}{object}

\item{TrimmingMethod}{TrimmingMethod}

\item{M1TrimmingCutoff}{M1TrimmingCutoff}

\item{M2CutoffQualityScore}{M2CutoffQualityScore}

\item{M2SlidingWindowSize}{M2SlidingWindowSize}
}
\description{
A SangerContig method which updates QualityReport parameter for each the SangerRead instance inside SangerContig.
}
\examples{
data("sangerContig")
updateQualityParam(sangerContig,
                   TrimmingMethod         = "M2",
                   M1TrimmingCutoff       = NULL,
                   M2CutoffQualityScore   = 40,
                   M2SlidingWindowSize    = 15)
}
