% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerContig.R
\docType{methods}
\name{SangerContig-class-writeFastaSC}
\alias{SangerContig-class-writeFastaSC}
\alias{writeFastaSC,SangerContig-method}
\title{writeFastaSC}
\usage{
\S4method{writeFastaSC}{SangerContig}(object, outputDir = NULL,
  compress = FALSE, compression_level = NA, selection = "all")
}
\description{
A SangerContig method which writes sequences into Fasta files.
}
\examples{
\dontrun{writeFastaSC(sangerContig, "/Users/chaokuan-hao/Desktop/sangeranalyseR_fasta/SangerContig")}
}
