% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangerContig.R
\docType{class}
\name{SangerContig-class}
\alias{SangerContig-class}
\title{SangerContig}
\description{
An S4 class for storing multiple single reads to build up new
 consensus read
}
\section{Slots}{

\describe{
\item{\code{consensusReadName}}{.}

\item{\code{suffixForwardRegExp}}{.}

\item{\code{suffixReverseRegExp}}{.}

\item{\code{secondaryPeakDF}}{}
}}

\examples{
rawDataDir <- system.file("extdata", package = "sangeranalyseR")
inputFilesParentDir <- file.path(rawDataDir, "Allolobophora_chlorotica")
consensusReadName <- "ACHLO006-09[LCO1490_t1,HCO2198_t1]"
suffixForwardRegExp <- "_[F]_[0-9]*.ab1"
suffixReverseRegExp <- "_[R]_[0-9]*.ab1"
A_chloroticConsensusReads <- new("SangerContig",
                                 parentDirectory       = inputFilesParentDir,
                                 consensusReadName    = consensusReadName,
                                 suffixForwardRegExp   = suffixForwardRegExp,
                                 suffixReverseRegExp   = suffixReverseRegExp,
                                 refAminoAcidSeq = "SRQWLFSTNHKDIGTLYFIFGAWAGMVGTSLSILIRAELGHPGALIGDDQIYNVIVTAHAFIMIFFMVMPIMIGGFGNWLVPLMLGAPDMAFPRMNNMSFWLLPPALSLLLVSSMVENGAGTGWTVYPPLSAGIAHGGASVDLAIFSLHLAGISSILGAVNFITTVINMRSTGISLDRMPLFVWSVVITALLLLLSLPVLAGAITMLLTDRNLNTSFFDPAGGGDPILYQHLFWFFGHPEVYILILPGFGMISHIISQESGKKETFGSLGMIYAMLAIGLLGFIVWAHHMFTVGMDVDTRAYFTSATMIIAVPTGIKIFSWLATLHGTQLSYSPAILWALGFVFLFTVGGLTGVVLANSSVDIILHDTYYVVAHFHYVLSMGAVFAIMAGFIHWYPLFTGLTLNNKWLKSHFIIMFIGVNLTFFPQHFLGLAGMPRRYSDYPDAYTTWNIVSTIGSTISLLGILFFFFIIWESLVSQRQVIYPIQLNSSIEWYQNTPPAEHSYSELPLLTN",
                                 TrimmingMethod        = "M2",
                                 M1TrimmingCutoff      = NULL,
                                 M2CutoffQualityScore  = 40,
                                 M2SlidingWindowSize   = 10,
                                 baseNumPerRow         = 100,
                                 heightPerRow          = 200,
                                 signalRatioCutoff     = 0.33,
                                 showTrimmed           = TRUE)
}
\author{
Kuan-Hao Chao
}
