% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangerContig.R
\docType{methods}
\name{SangerContig}
\alias{SangerContig}
\title{Constructor method of SangerContig Class.}
\usage{
\S4method{initialize}{SangerContig}(.Object, ..., inputSource = "ABIF",
  fastaFileName = NULL, namesConversionCSV = NULL,
  parentDirectory = NULL, contigName = "",
  suffixForwardRegExp = "_F.ab1", suffixReverseRegExp = "_R.ab1",
  geneticCode = GENETIC_CODE, TrimmingMethod = "M1",
  M1TrimmingCutoff = 1e-04, M2CutoffQualityScore = NULL,
  M2SlidingWindowSize = NULL, baseNumPerRow = 100,
  heightPerRow = 200, signalRatioCutoff = 0.33, showTrimmed = TRUE,
  refAminoAcidSeq = "", minReadsNum = 2, minReadLength = 20,
  minFractionCall = 0.5, maxFractionLost = 0.5,
  acceptStopCodons = TRUE, readingFrame = 1, processorsNum = NULL)
}
\description{
Constructor method of SangerContig Class.
}
