% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\docType{methods}
\name{SangerRead-class-generateReportSR}
\alias{SangerRead-class-generateReportSR}
\alias{generateReportSR,SangerRead-method}
\title{generateReportSR}
\usage{
\S4method{generateReportSR}{SangerRead}(object, outputDir,
  navigationContigFN = NULL, navigationAlignmentFN = NULL)
}
\arguments{
\item{object}{object}

\item{outputDir}{outputDir}

\item{navigationContigFN}{navigationContigFN}

\item{navigationAlignmentFN}{navigationAlignmentFN}
}
\description{
A SangerRead method which generates final reports of the SangerRead instance.
}
\examples{
\dontrun{generateReportSR(sangerReadF)}
}
