% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\docType{methods}
\name{SangerRead-class-generateReportSR}
\alias{SangerRead-class-generateReportSR}
\title{generateReportSR}
\usage{
\S4method{generateReportSR}{SangerRead}(obj, outputDir,
  navigationContigFN = NULL, navigationAlignmentFN = NULL)
}
\description{
generateReportSR
}
\examples{
load("data/sangerRead.RData")
generateReportSR(sangerRead)
}
