% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\docType{methods}
\name{SangerRead-class-generateReportSR}
\alias{SangerRead-class-generateReportSR}
\alias{generateReportSR,SangerRead-method}
\title{generateReportSR}
\usage{
\S4method{generateReportSR}{SangerRead}(object, outputDir,
  navigationContigFN = NULL, navigationAlignmentFN = NULL)
}
\arguments{
\item{object}{A SangerRead S4 instance.}

\item{outputDir}{The output directory of the generated HTML report.}

\item{navigationContigFN}{The internal parameter passed to HTML report. Users should not modify this parameter on their own.}

\item{navigationAlignmentFN}{The internal parameter passed to HTML report. Users should not modify this parameter on their own.}
}
\description{
A SangerRead method which generates final reports of the SangerRead instance.
}
\examples{
data("sangerReadFData")
\dontrun{
generateReportSR(sangerReadFData)}
}
