% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\docType{methods}
\name{SangerRead-class-MakeBaseCalls}
\alias{SangerRead-class-MakeBaseCalls}
\alias{MakeBaseCalls,SangerRead-method}
\title{MakeBaseCalls}
\usage{
\S4method{MakeBaseCalls}{SangerRead}(object, signalRatioCutoff = 0.33)
}
\description{
A SangerRead method which does base calling on SangerRead instance
}
\examples{
\dontrun{MakeBaseCalls(sangerReadF, signalRatioCutoff = 0.22)}
}
