% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\docType{methods}
\name{SangerRead-class-MakeBaseCalls}
\alias{SangerRead-class-MakeBaseCalls}
\title{MakeBaseCalls}
\usage{
\S4method{MakeBaseCalls}{SangerRead}(obj, signalRatioCutoff = 0.33)
}
\description{
MakeBaseCalls
}
\examples{
load("data/sangerRead.RData")
MakeBaseCalls(sangerReadF, signalRatioCutoff = 0.22)
}
