% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\docType{methods}
\name{SangerRead-class-qualityBasePlot}
\alias{SangerRead-class-qualityBasePlot}
\title{qualityBasePlot}
\usage{
\S4method{qualityBasePlot}{SangerRead}(object)
}
\description{
qualityBasePlot
}
\examples{
load("data/sangerRead.RData")
qualityBasePlot(sangerReadF)
}
