% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\docType{methods}
\name{SangerRead-class-qualityBasePlot}
\alias{SangerRead-class-qualityBasePlot}
\alias{qualityBasePlot,SangerRead-method}
\title{qualityBasePlot}
\usage{
\S4method{qualityBasePlot}{SangerRead}(object)
}
\arguments{
\item{object}{A SangerRead S4 instance.}
}
\value{
A quality plot.
}
\description{
A SangerRead method which creates quality base interactive plot.
}
\examples{
data("sangerReadFData")
\dontrun{
qualityBasePlot(sangerReadFData)}
}
