% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\name{SangerRead-class-readTable}
\alias{SangerRead-class-readTable}
\alias{readTable,SangerRead-method}
\title{readTable}
\usage{
\S4method{readTable}{SangerRead}(object, indentation = 0)
}
\arguments{
\item{object}{A SangerRead S4 instance.}

\item{indentation}{The indentation for different level printing.}
}
\value{
None
}
\description{
A SangerRead method which generates summary table for SangerRead instance
}
\examples{
data(sangerReadFData)
data(sangerContigData)
\donttest{
readTable(sangerReadFData)
readTable(sangerContigData)
}
}
