% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\docType{methods}
\name{SangerRead-class-readTable}
\alias{SangerRead-class-readTable}
\alias{readTable,readTable-method}
\title{readTable}
\usage{
\S4method{readTable}{SangerRead}(object)
}
\arguments{
\item{object}{A SangerRead S4 instance.}
}
\value{
None
}
\description{
A SangerRead method which generates summary table for SangerRead instance
}
\examples{
data(sangerReadFData)
data(sangerContigData)
data(sangerAlignmentData)
\dontrun{
readTable(sangerReadFData)
readTable(sangerContigData)
readTable(sangerAlignmentData)
}
}
