% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\docType{methods}
\name{SangerRead-class-updateQualityParam}
\alias{SangerRead-class-updateQualityParam}
\title{updateQualityParam}
\usage{
\S4method{updateQualityParam}{SangerRead}(object, TrimmingMethod = "M1",
  M1TrimmingCutoff = 1e-04, M2CutoffQualityScore = NULL,
  M2SlidingWindowSize = NULL)
}
\description{
updateQualityParam
}
\examples{
\dontrun{
load("data/sangerRead.RData")
updateQualityParam(sangerReadF,
                   TrimmingMethod         = "M2",
                   M1TrimmingCutoff       = NULL,
                   M2CutoffQualityScore   = 40,
                   M2SlidingWindowSize    = 15)
}
}
