% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\docType{methods}
\name{SangerRead-class-writeFastaSR}
\alias{SangerRead-class-writeFastaSR}
\alias{writeFastaSR,SangerRead-method}
\title{writeFastaSR}
\usage{
\S4method{writeFastaSR}{SangerRead}(object, outputDir = NULL,
  compress = FALSE, compression_level = NA)
}
\arguments{
\item{object}{object}

\item{outputDir}{outputDir}

\item{compress}{compress}

\item{compression_level}{compression_level}
}
\description{
A SangerRead method which writes the sequence into Fasta files.
}
\examples{
data("sangerReadF")
writeFastaSR(sangerReadF)
}
