% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\docType{methods}
\name{SangerRead-class-writeFastaSR}
\alias{SangerRead-class-writeFastaSR}
\title{writeFastaSR}
\usage{
\S4method{writeFastaSR}{SangerRead}(obj, outputDir = NULL,
  compress = FALSE, compression_level = NA)
}
\description{
writeFastaSR
}
\examples{
load("data/sangerRead.RData")
writeFastaSR(sangerReadF, "/Users/chaokuan-hao/Desktop/sangeranalyseR_fasta/SangerRead")
}
