% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangerRead.R
\docType{class}
\name{SangerRead-class}
\alias{SangerRead-class}
\title{SangerRead}
\description{
An S4 class extending sangerseq S4 class which corresponds to a single ABIF file in Sanger sequencing.
}
\section{Slots}{

\describe{
\item{\code{inputSource}}{The input source of the raw file. It must be \code{"ABIF"} or \code{"FASTA"}. The default value is \code{"ABIF"}.}

\item{\code{readFeature}}{The direction of the Sanger read. The value must be \code{"Forward Read"} or \code{"Reverse Read"}.}

\item{\code{readFileName}}{The filename of the target ABIF file.}

\item{\code{abifRawData}}{A S4 class containing all fields in the ABIF file. It is defined in sangerseqR package.}

\item{\code{QualityReport}}{A S4 class containing quality trimming related inputs and trimming results.}

\item{\code{ChromatogramParam}}{A S4 class containing chromatogram inputs.}

\item{\code{primaryAASeqS1}}{A polypeptide translated from primary DNA sequence starting from the first nucleic acid.}

\item{\code{primaryAASeqS2}}{A polypeptide translated from primary DNA sequence starting from the second nucleic acid.}

\item{\code{primaryAASeqS3}}{A polypeptide translated from primary DNA sequence starting from the third nucleic acid.}

\item{\code{geneticCode}}{Named character vector in the same format as \code{GENETIC_CODE} (the default), which represents the standard genetic code. This is the code with which the function will attempt to translate your DNA sequences. You can get an appropriate vector with the getGeneticCode() function. The default is the standard code.}

\item{\code{primarySeqRaw}}{The raw primary sequence from sangerseq class in sangerseqR package before base calling.}

\item{\code{secondarySeqRaw}}{The raw secondary sequence from sangerseq class in sangerseqR package before base calling.}

\item{\code{peakPosMatrixRaw}}{The raw peak position matrix from sangerseq class in sangerseqR package before base calling.}

\item{\code{peakAmpMatrixRaw}}{The raw peak amplitude matrix from sangerseq class in sangerseqR package before base calling.}
}}

\examples{
## Input From ABIF file format
# Forward Read
inputFilesPath <- system.file("extdata/", package = "sangeranalyseR")
A_chloroticaFFN <- file.path(inputFilesPath,
                             "Allolobophora_chlorotica",
                             "ACHLO",
                             "ACHLO006-09[LCO1490_t1,HCO2198_t1]_F_1.ab1")
sangerReadF <- new("SangerRead",
                    inputSource           = "ABIF",
                    readFeature           = "Forward Read",
                    readFileName          = A_chloroticaFFN,
                    geneticCode           = GENETIC_CODE,
                    TrimmingMethod        = "M1",
                    M1TrimmingCutoff      = 0.0001,
                    M2CutoffQualityScore  = NULL,
                    M2SlidingWindowSize   = NULL,
                    baseNumPerRow         = 100,
                    heightPerRow          = 200,
                    signalRatioCutoff     = 0.33,
                    showTrimmed           = TRUE)

# Reverse Read
A_chloroticaRFN <- file.path(inputFilesPath,
                             "Allolobophora_chlorotica",
                             "ACHLO",
                             "ACHLO006-09[LCO1490_t1,HCO2198_t1]_R_2.ab1")
sangerReadR <- new("SangerRead",
                    inputSource           = "ABIF",
                    readFeature           = "Reverse Read",
                    readFileName          = A_chloroticaRFN,
                    geneticCode           = GENETIC_CODE,
                    TrimmingMethod        = "M1",
                    M1TrimmingCutoff      = 0.0001,
                    M2CutoffQualityScore  = NULL,
                    M2SlidingWindowSize   = NULL,
                    baseNumPerRow         = 100,
                    heightPerRow          = 200,
                    signalRatioCutoff     = 0.33,
                    showTrimmed           = TRUE)

## Input From FASTA file format
# Forward Read
inputFilesPath <- system.file("extdata/", package = "sangeranalyseR")
A_chloroticaFFNfa <- file.path(inputFilesPath,
                               "fasta",
                               "SangerRead",
                               "ACHLO006-09[LCO1490_t1,HCO2198_t1]_F_1.fa")
readNameFfa <- "ACHLO006-09[LCO1490_t1,HCO2198_t1]_F_1"
sangerReadFfa <- new("SangerRead",
                     inputSource   = "FASTA",
                     readFeature   = "Forward Read",
                     readFileName  = A_chloroticaFFNfa,
                     fastaReadName = readNameFfa,
                     geneticCode   = GENETIC_CODE)
# Reverse Read
A_chloroticaRFNfa <- file.path(inputFilesPath,
                               "fasta",
                               "SangerRead",
                               "ACHLO006-09[LCO1490_t1,HCO2198_t1]_R_2.fa")
readNameRfa <- "ACHLO006-09[LCO1490_t1,HCO2198_t1]_R_2"
sangerReadRfa <- new("SangerRead",
                     inputSource   = "FASTA",
                     readFeature   = "Reverse Read",
                     readFileName  = A_chloroticaRFNfa,
                     fastaReadName = readNameRfa,
                     geneticCode   = GENETIC_CODE)
}
\author{
Kuan-Hao Chao
}
