% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangerRead.R
\name{SangerRead-class}
\alias{SangerRead-class}
\alias{SangerRead}
\title{SangerRead}
\usage{
SangerRead(readFeature = character(0), readFileName = character(0),
  geneticCode = GENETIC_CODE, TrimmingMethod = "M1",
  M1TrimmingCutoff = 1e-04, M2CutoffQualityScore = NULL,
  M2SlidingWindowSize = NULL, baseNumPerRow = 100,
  heightPerRow = 200, signalRatioCutoff = 0.33, showTrimmed = TRUE)
}
\description{
An S4 class extending sangerseq S4 class
}
\section{Slots}{

\describe{
\item{\code{readFeature}}{.}

\item{\code{readFeature}}{.}

\item{\code{readFileName}}{.}

\item{\code{abifRawData}}{.}

\item{\code{QualityReport}}{.}

\item{\code{ChromatogramParam}}{.}

\item{\code{primaryAASeq}}{.}

\item{\code{geneticCode}}{.}

\item{\code{primarySeqRaw}}{.}

\item{\code{secondarySeqRaw}}{.}

\item{\code{peakPosMatrixRaw}}{.}

\item{\code{peakAmpMatrixRaw}}{.}
}}

\examples{
inputFilesPath <- system.file("extdata/", package = "sangeranalyseR")
A_chloroticaFdReadFN <- file.path(inputFilesPath,
                                  "Allolobophora_chlorotica",
                                  "ACHLO006-09[LCO1490_t1,HCO2198_t1]_F_1.ab1")
A_chloroticaRead <- new("SangerRead",
                        readFeature           = "Forward Read",
                        readFileName          = A_chloroticaFdReadFN,
                        geneticCode           = GENETIC_CODE,
                        TrimmingMethod        = "M1",
                        M1TrimmingCutoff      = 0.0001,
                        M2CutoffQualityScore  = NULL,
                        M2SlidingWindowSize   = NULL,
                        baseNumPerRow         = 100,
                        heightPerRow          = 200,
                        signalRatioCutoff     = 0.33,
                        showTrimmed           = TRUE)
}
\author{
Kuan-Hao Chao
}
