% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodSangerRead.R
\docType{methods}
\name{SangerRead-class-qualityBasePlot}
\alias{SangerRead-class-qualityBasePlot}
\alias{qualityBasePlot,SangerRead-method}
\alias{SangerRead-class-updateQualityParam}
\alias{updateQualityParam,SangerRead-method}
\alias{SangerRead-class-MakeBaseCalls}
\alias{MakeBaseCalls,SangerRead-method}
\alias{SangerRead-class-writeFastaSR}
\alias{writeFastaSR,SangerRead-method}
\alias{SangerRead-class-generateReportSR}
\alias{generateReportSR,SangerRead-method}
\title{qualityBasePlot}
\usage{
\S4method{qualityBasePlot}{SangerRead}(object)

\S4method{updateQualityParam}{SangerRead}(object, TrimmingMethod = "M1",
  M1TrimmingCutoff = 1e-04, M2CutoffQualityScore = NULL,
  M2SlidingWindowSize = NULL)

\S4method{MakeBaseCalls}{SangerRead}(object, signalRatioCutoff = 0.33)

\S4method{writeFastaSR}{SangerRead}(object, outputDir = NULL,
  compress = FALSE, compression_level = NA)

\S4method{generateReportSR}{SangerRead}(object, outputDir,
  navigationContigFN = NULL, navigationAlignmentFN = NULL)
}
\arguments{
\item{object}{.}

\item{TrimmingMethod}{.}

\item{M1TrimmingCutoff}{.}

\item{M2CutoffQualityScore}{.}

\item{M2SlidingWindowSize}{.}

\item{signalRatioCutoff}{.}

\item{outputDir}{.}

\item{compress}{.}

\item{compression_level}{.}

\item{navigationContigFN}{.}

\item{navigationAlignmentFN}{.}

\item{object}{.}

\item{object}{.}

\item{object}{.}

\item{object}{.}

\item{outputDir}{.}
}
\description{
A SangerRead method which creates quality base interactive plot.

A SangerRead method which updates QualityReport parameter inside the SangerRead.

A SangerRead method which does base calling on SangerRead instance

A SangerRead method which writes the sequence into Fasta files.

A SangerRead method which generates final reports of the SangerRead instance.
}
\examples{
\dontrun{qualityBasePlot(sangerReadF)}
\dontrun{updateQualityParam(sangerReadF,
                   TrimmingMethod         = "M2",
                   M1TrimmingCutoff       = NULL,
                   M2CutoffQualityScore   = 40,
                   M2SlidingWindowSize    = 15)}
\dontrun{MakeBaseCalls(sangerReadF, signalRatioCutoff = 0.22)}
\dontrun{writeFastaSR(sangerReadF, "/Users/chaokuan-hao/Desktop/sangeranalyseR_fasta/SangerRead")}
\dontrun{generateReportSR(sangerReadF)}
}
