% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangerRead.R
\docType{methods}
\name{SangerRead}
\alias{SangerRead}
\title{Constructor method of SangerRead Class.}
\usage{
\S4method{initialize}{SangerRead}(.Object, ..., inputSource = "ABIF",
  readFeature = "", readFileName = "", fastaReadName = NULL,
  namesConversionCSV = NULL, geneticCode = GENETIC_CODE,
  TrimmingMethod = "M1", M1TrimmingCutoff = 1e-04,
  M2CutoffQualityScore = NULL, M2SlidingWindowSize = NULL,
  baseNumPerRow = 100, heightPerRow = 200, signalRatioCutoff = 0.33,
  showTrimmed = TRUE)
}
\description{
Constructor method of SangerRead Class.
}
