% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SangerReads-class}
\alias{SangerReads-class}
\title{SangerReads}
\description{
An S4 class for storing reads of sanger sequencing.
}
\section{Slots}{

\describe{
\item{\code{forward.read}}{.}

\item{\code{reverse.read}}{.}

\item{\code{consensus.read}}{.}

\item{\code{quality.matrix}}{.}
}}

\examples{
inputFilesPath <- system.file("extdata/", package = "sangeranalyseR")
A_chloroticaForwardReadFN <- file.path(inputFilesPath,
                                       "Allolobophora_chlorotica",
                                       "ACHLO006-09[LCO1490_t1,HCO2198_t1]_F.ab1")
A_chloroticaReverseReadFN <- file.path(inputFilesPath,
                                       "Allolobophora_chlorotica",
                                       "ACHLO006-09[LCO1490_t1,HCO2198_t1]_R.ab1")
A_chloroticaRead <- new("SangerReads",
                         forwardReadFileName = A_chloroticaForwardReadFN,
                         reverseReadFileName = A_chloroticaReverseReadFN)
}
\author{
Kuan-Hao Chao
}
