% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangerSingleRead.R
\name{SangerSingleRead-class}
\alias{SangerSingleRead-class}
\alias{SangerSingleRead}
\title{SangerSingleRead}
\usage{
SangerSingleRead(readFeature = character(0),
  readFileName = character(0), cutoffQualityScore = 20L,
  slidingWindowSize = 5L)
}
\description{
An S4 class extending sangerseq S4 class
}
\section{Slots}{

\describe{
\item{\code{readFeature}}{.}
}}

\examples{
inputFilesPath <- system.file("extdata/", package = "sangeranalyseR")
A_chloroticaFdReadFN <- file.path(inputFilesPath,
                                  "Allolobophora_chlorotica",
                                  "ACHLO006-09[LCO1490_t1,HCO2198_t1]_F.ab1")
A_chloroticaSingleRead <- new("SangerSingleRead",
                              readFeature         = "ForwardRead",
                              readFileName        = A_chloroticaFdReadFN,
                              cutoffQualityScore  = 60L,
                              slidingWindowSize   = 8L)
}
\author{
Kuan-Hao Chao
}
