% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassSangerSingleRead.R
\name{SangerSingleRead-class}
\alias{SangerSingleRead-class}
\alias{SangerSingleRead}
\title{SangerSingleRead}
\usage{
SangerSingleRead(readFeature = character(0),
  readFileName = character(0), TrimmingMethod = "M2",
  M1TrimmingCutoff = NULL, M2CutoffQualityScore = 40,
  M2SlidingWindowSize = 10)
}
\description{
An S4 class extending sangerseq S4 class
}
\section{Slots}{

\describe{
\item{\code{readFeature}}{.}

\item{\code{readFeature}}{.}

\item{\code{readFileName}}{.}

\item{\code{abifRawData}}{.}

\item{\code{QualityReport}}{.}

\item{\code{ChromatogramParam}}{.}

\item{\code{primaryAASeq}}{.}
}}

\examples{
inputFilesPath <- system.file("extdata/", package = "sangeranalyseR")
A_chloroticaFdReadFN <- file.path(inputFilesPath,
                                  "Allolobophora_chlorotica",
                                  "RBNII396-13[C_LepFolF,C_LepFolR]_F_1.ab1")
A_chloroticaSingleRead <- new("SangerSingleRead",
                              readFeature           = "ForwardRead",
                              readFileName          = A_chloroticaFdReadFN,
                              TrimmingMethod        = "M2",
                              M1TrimmingCutoff      = NULL,
                              M2CutoffQualityScore  = 40,
                              M2SlidingWindowSize   = 10,
                              baseNumPerRow         = 100,
                              signalRatioCutoff     = 0.33,
                              showTrimmed           = TRUE)
}
\author{
Kuan-Hao Chao
}
