% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.abi.file.R
\name{summarise.abi.file}
\alias{summarise.abi.file}
\title{Create a detailed summary of a single ABI sequencing file}
\usage{
summarise.abi.file(inputfile, trim.cutoff = 1e-04,
  secondary.peak.cutoff = 0.33, write.secondary.peak.files = FALSE,
  processors = NULL)
}
\arguments{
\item{processors}{The number of processors to use, or NULL (the default) for all available processors}

\item{input.folder}{the folder in which to search recursively for *.ab1 files.}

\item{cutoff}{the ratio of the height of a secondary peak to a primary peak. Secondary peaks higher than this ratio are annotated. Those below the ratio are not.}

\item{write.files}{TRUE/FALSE. FALSE (the default): do not write any files to disk. TRUE: write plots and .csv files of secondary peaks to disk, in this case, files are named with the same filename as the .ab1 files, but with .csv and .pdf extensions, and are saved in the same folder as the .ab1 file from which they were created.}
}
\value{
output a dataframe of secondary peaks
}
\description{
This function recursively scans a folder that contains abi files either in itself or its subfolders, and looks for secondary peaks in each sequence. It produces an annotated plot of the secondary peaks above a user-defined cutoff, as well as a table of the mismatches.
}
\details{
It recursively scanes the input folder for all abi files. For each abi file, it creates and returns a data frame and a plot of mismatches, and also writes both to the same file as the .ab1 file (as a .csv and .pdf respectively).
}
\keyword{chromatogram,}
\keyword{mismatch}
\keyword{peak,}

