% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.merged.read.R
\name{summarise.merged.read}
\alias{summarise.merged.read}
\title{Produce a one-row summary of a merged read}
\usage{
summarise.merged.read(merged.read)
}
\arguments{
\item{merged.read}{a merged.read object produced by the merge.reads() function}
}
\value{
A one-row summary of the merged read
}
\description{
Produce a one-row summary of a merged read
}

