% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.to.reference.R
\name{trim.to.reference}
\alias{trim.to.reference}
\title{Trim an alignment to the limits of a supplied reference sequence}
\usage{
trim.to.reference(alignment, reference)
}
\arguments{
\item{alignment}{a DNAStringSet object}

\item{reference}{a DNA sequence as a DNAString sequence}
}
\description{
This function profile aligns the alignment to the
reference sequence. It then trims the alignment to the limits of the 
reference sequence, and returns the trimmed alignment.
}

