% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{updateQualityParam}
\alias{updateQualityParam}
\title{Method updateQualityParam}
\usage{
updateQualityParam(object, TrimmingMethod = "M1",
  M1TrimmingCutoff = 1e-04, M2CutoffQualityScore = NULL,
  M2SlidingWindowSize = NULL)
}
\description{
Method updateQualityParam
}
