% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{updateQualityParam}
\alias{updateQualityParam}
\title{Method updateQualityParam}
\usage{
updateQualityParam(object, TrimmingMethod = "M1",
  M1TrimmingCutoff = 1e-04, M2CutoffQualityScore = NULL,
  M2SlidingWindowSize = NULL)
}
\arguments{
\item{object}{object}

\item{TrimmingMethod}{TrimmingMethod}

\item{M1TrimmingCutoff}{M1TrimmingCutoff}

\item{M2CutoffQualityScore}{M2CutoffQualityScore}

\item{M2SlidingWindowSize}{M2SlidingWindowSize}
}
\description{
Method updateQualityParam
}
