% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodShared.R
\name{writeFasta}
\alias{writeFasta}
\title{Method writeFasta}
\usage{
writeFasta(object, outputDir = NULL, compress = FALSE,
  compression_level = NA, selection = "all")
}
\arguments{
\item{object}{object}

\item{outputDir}{outputDir}

\item{compress}{compress}

\item{compression_level}{compression_level}

\item{selection}{selection}
}
\description{
A method which writes FASTA files of the SangerRead,
 SangerContig, and SangerAlignment instance.
}
\examples{
data(sangerReadF)
data(sangerContig)
data(sangerAlignment)
\dontrun{
writeFasta(sangerReadF)
writeFasta(sangerContig)
writeFasta(sangerAlignment)}
}
\author{
Kuan-Hao Chao
}
