% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{writeFastaSA}
\alias{writeFastaSA}
\title{Method writeFastaSA}
\usage{
writeFastaSA(object, outputDir = NULL, compress = FALSE,
  compression_level = NA, selection = "all")
}
\description{
Method writeFastaSA
}
