% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{writeFastaSC}
\alias{writeFastaSC}
\title{Method writeFastaSC}
\usage{
writeFastaSC(object, outputDir = NULL, compress = FALSE,
  compression_level = NA, selection = "all")
}
\arguments{
\item{object}{object}

\item{outputDir}{outputDir}

\item{compress}{compress}

\item{compression_level}{compression_level}

\item{selection}{selection}
}
\description{
Method writeFastaSC
}
\examples{
data(sangerContig)
writeFastaSC(sangerContig)
}
