% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{writeFastaSC}
\alias{writeFastaSC}
\title{Method writeFastaSC}
\usage{
writeFastaSC(obj, outputDir = NULL, compress = FALSE,
  compression_level = NA, selection = "all")
}
\description{
Method writeFastaSC
}
