% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{writeFastaSR}
\alias{writeFastaSR}
\title{Method writeFastaSR}
\usage{
writeFastaSR(
  object,
  outputDir = NULL,
  compress = FALSE,
  compression_level = NA
)
}
\arguments{
\item{object}{A SangerRead S4 instance.}

\item{outputDir}{The output directory of the generated FASTA file.}

\item{compress}{Like for the \code{save} function in base R, must be \code{TRUE} or \code{FALSE} (the default), or a single string specifying whether writing to the file is to use compression. The only type of compression supported at the moment is "gzip". This parameter will be passed to \code{writeXStringSet} function in Biostrings package.}

\item{compression_level}{This parameter will be passed to \code{writeXStringSet} function in Biostrings package.}
}
\value{
The output absolute path to the FASTA file.
}
\description{
Method writeFastaSR
}
\examples{
data(sangerReadFData)
writeFastaSR(sangerReadFData)
}
